refactor(resolving-deps-resolver): split resolve_peers into focused modules
resolve_peers.rs carried the public entry points, workspace orchestration,
peer-hoist discovery, the walker, cache matching, child realization, final
dep-path construction, cycle analysis, graph construction, and the standalone
peer-context utilities in one 3,934-line file, with a 2,434-line test module
beside it.

Keep resolve_peers.rs as the facade holding the options/result types and the
two public entry points, and move the implementation into resolve_peers/:
discovery.rs (the hoist-loop discovery engine and its persistent caches),
walker.rs (Walker, its state, walk, resolve_node, and peer resolution),
cache.rs (purePkgs and peersCache matching, deferred child resolution, child
realization, retention decisions), finalize.rs (pending-edge repair, SCC cycle
analysis, final depPath recomputation, graph construction), and context.rs
(parent references, shared ancestor chains, link remapping, peer-suffix
parsing, range matching).

Walker stays one struct; its inherent impl is divided across the modules with
`pub(super)` visibility, so every cross-module item stays private to
resolve_peers and the crate's public API is unchanged. The unit tests that
drive one module's internals move next to it, with the shared tree builders
lifted into a test_support module; the end-to-end peer-resolution tests stay
at the facade.

Pure code movement: no behavior, allocation, cloning, dispatch, or traversal
change, and release builds use `lto = "fat"` with `codegen-units = 1`, so the
module boundaries cannot shift inlining either.

Closes pnpm/pnpm#13548.
