+++ pnpm/crates/resolving-deps-resolver/src/resolve_peers.rs
@@ -1,7 +1,7 @@
 //! Resolves peer dependencies for a resolved dependency tree.
 //!
 //! Walks the per-occurrence [`crate::ResolvedTree::dependencies_tree`]
-//! depth-first, propagating a [`ParentRefs`] map of available parents
+//! depth-first, propagating a [`ParentRefs`](context::ParentRefs) map of available parents
 //! down the chain, and matches each visited package's
 //! [`crate::ResolvedPackage::peer_dependencies`] against that map.
 //! Produces a [`DependenciesGraph`] keyed by depPath plus the
@@ -30,102 +30,32 @@
 //! `in_progress` set, where a re-entry on the same `NodeId` falls back
 //! to `name@version` as the peer-id.
 
+mod cache;
+mod context;
+mod discovery;
+mod finalize;
+mod walker;
+
 use crate::{
     dedupe_injected_deps::dedupe_injected_deps,
     dedupe_peer_dependents::dedupe_peer_dependents,
-    dependencies_graph::{
-        DependenciesGraph, DependenciesGraphNode, MissingPeer, ParentPackageRef,
-        PeerDependencyIssue, PeerDependencyIssues,
-    },
+    dependencies_graph::{DependenciesGraph, PeerDependencyIssues},
     node_id::NodeId,
-    resolved_tree::{
-        AncestorIds, ChildEdge, DependenciesTreeNode, DirectDep, PeerDep, ResolvedPackage,
-        ResolvedTree, TreeChildren,
-    },
+    resolved_tree::{DirectDep, ResolvedTree},
 };
-use node_semver::{Range, Version};
-
-use pacquet_deps_path::{
-    DepPath, PeerId, create_peer_dep_graph_hash, index_of_dep_path_suffix,
-    link_path_to_peer_version,
+use context::{
+    CurrentProviderSource, SharedChain, importer_relative_link_dep_path, node_id_sort_key,
 };
-use pacquet_resolving_resolver_base::{ResolveResult, get_peer_version_range};
+use discovery::PeerDiscoveryCaches;
+pub(crate) use discovery::{PeerDiscoveryResult, PeerHoistDiscovery, apply_hoist_missing_scope};
+use pacquet_deps_path::DepPath;
 use rustc_hash::{FxHashMap as HashMap, FxHashSet as HashSet};
 use std::{
-    collections::{BTreeMap, BTreeSet},
+    collections::BTreeMap,
     path::{Path, PathBuf},
     sync::Arc,
 };
-
-/// Pull `(name, version)` out of a `ResolveResult` the peer-resolution
-/// stage can hash and compare on.
-///
-/// The npm-registry resolver always fills [`ResolveResult::name_ver`],
-/// so the fast path lifts it straight out. The git / tarball / local
-/// resolvers leave it `None` (their canonical name lives in the
-/// fetched manifest, which the resolver doesn't read at resolve
-/// time); for those, fall back to `(alias, id-as-string)`. The peer
-/// graph machinery only ever looks the name up in
-/// [`ResolvedTree::all_peer_dep_names`] — a set built by parsing the
-/// peer dependencies of npm-shaped packages — so the fallback's
-/// "name" will simply miss every lookup, naturally short-circuiting
-/// peer propagation for non-npm packages without panicking on
-/// `name_ver = None`.
-/// Reinterpret a `link:<rel>` [`NodeId`] as a [`DepPath`].
-///
-/// Linked top-parent `NodeIds` (whether the workspace-link arm or the
-/// `excludeLinksFromLockfile` remap) never enter the dependency tree,
-/// so [`Walker::node_dep_paths`] never maps them. The `link:<rel>`
-/// `NodeId` is itself a well-formed pnpm `DepPath`, so the snapshot
-/// child edge can use it verbatim.
-fn link_node_id_as_dep_path(node_id: &NodeId) -> Option<DepPath> {
-    let NodeId::Leaf(id) = node_id else { return None };
-    id.starts_with("link:").then(|| DepPath::from(id.to_string()))
-}
-
-fn importer_relative_link_dep_path(
-    dep_path: &DepPath,
-    lockfile_dir: Option<&Path>,
-    project_dir: Option<&Path>,
-) -> DepPath {
-    let Some(target) = dep_path.as_str().strip_prefix("link:") else {
-        return dep_path.clone();
-    };
-    let (Some(lockfile_dir), Some(project_dir)) = (lockfile_dir, project_dir) else {
-        return dep_path.clone();
-    };
-    let target = Path::new(target);
-    let absolute_target = if target.is_absolute() {
-        pacquet_fs::lexical_normalize(target)
-    } else {
-        pacquet_fs::lexical_normalize(&lockfile_dir.join(target))
-    };
-    // `diff_paths` walks both paths component-wise, so a base still
-    // carrying `.` / `..` segments would consume them as real directories
-    // and count the wrong number of `..` hops back out.
-    let project_dir = pacquet_fs::lexical_normalize(project_dir);
-    let relative_target = pathdiff::diff_paths(&absolute_target, project_dir)
-        .unwrap_or(absolute_target)
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    DepPath::from(format!("link:{relative_target}"))
-}
-
-fn node_id_sort_key(node_id: &NodeId) -> String {
-    match node_id {
-        NodeId::Counter(value) => format!("0:{value:020}"),
-        NodeId::Leaf(value) => format!("1:{value}"),
-    }
-}
-
-fn pkg_name_version(result: &ResolveResult) -> (String, String) {
-    if let Some(name_ver) = result.name_ver.as_ref() {
-        return (name_ver.name.to_string(), name_ver.suffix.to_string());
-    }
-    let fallback_name = result.alias.clone().unwrap_or_else(|| result.id.as_str().to_string());
-    (fallback_name, result.id.as_str().to_string())
-}
+use walker::Walker;
 
 /// Options threaded into [`fn@resolve_peers`].
 #[derive(Debug, Clone)]
@@ -219,6 +149,26 @@ pub struct ResolvePeersOptions {
     pub explicitly_requested_direct_dependencies: HashSet<String>,
 }
 
+impl Default for ResolvePeersOptions {
+    fn default() -> Self {
+        ResolvePeersOptions {
+            peers_suffix_max_length: 1000,
+            dedupe_peers: false,
+            exclude_links_from_lockfile: false,
+            lockfile_dir: None,
+            project_dir: None,
+            modules_dir: None,
+            hoist_missing_scope: None,
+            hoisted_optional_peer_node_ids: HashSet::default(),
+            hoisted_peer_provider_node_ids: HashSet::default(),
+            resolved_peer_provider_paths: None,
+            collect_paths_by_node_id: false,
+            declared_direct_dependencies: HashSet::default(),
+            explicitly_requested_direct_dependencies: HashSet::default(),
+        }
+    }
+}
+
 /// See [`ResolvePeersOptions::hoist_missing_scope`].
 #[derive(Debug, Clone)]
 pub struct HoistMissingScope {
@@ -267,26 +217,6 @@ impl HoistMissingScope {
     }
 }
 
-impl Default for ResolvePeersOptions {
-    fn default() -> Self {
-        ResolvePeersOptions {
-            peers_suffix_max_length: 1000,
-            dedupe_peers: false,
-            exclude_links_from_lockfile: false,
-            lockfile_dir: None,
-            project_dir: None,
-            modules_dir: None,
-            hoist_missing_scope: None,
-            hoisted_optional_peer_node_ids: HashSet::default(),
-            hoisted_peer_provider_node_ids: HashSet::default(),
-            resolved_peer_provider_paths: None,
-            collect_paths_by_node_id: false,
-            declared_direct_dependencies: HashSet::default(),
-            explicitly_requested_direct_dependencies: HashSet::default(),
-        }
-    }
-}
-
 /// Output bag of [`fn@resolve_peers`].
 #[derive(Debug, Default)]
 pub struct ResolvePeersResult {
@@ -308,203 +238,6 @@ pub struct ResolvePeersResult {
     /// [`ResolvePeersOptions::resolved_peer_provider_paths`] to run a
     /// locked-peer-provider reuse pass.
     pub paths_by_node_id: HashMap<NodeId, DepPath>,
-}
-
-/// What one peer-hoist discovery pass reports back to the hoist loop —
-/// the subset of [`ResolvePeersResult`] the loop actually consumes, so
-/// discovery never has to build a [`DependenciesGraph`].
-#[derive(Debug, Default)]
-pub(crate) struct PeerDiscoveryResult {
-    /// See [`ResolvePeersResult::resolved_peer_providers_by_alias`].
-    pub(crate) resolved_peer_providers_by_alias: BTreeMap<String, NodeId>,
-    pub(crate) peer_dependency_issues: PeerDependencyIssues,
-    /// Ancestor `pkgIdWithPatchHash` chains recorded per missing-peer
-    /// issue, consumed by [`fn@apply_hoist_missing_scope`].
-    missing_ancestor_pkg_ids: HashMap<String, Vec<SharedChain<String>>>,
-    /// See [`ResolvePeersResult::missing_names_by_pkg`].
-    pub(crate) missing_names_by_pkg: HashMap<String, HashSet<String>>,
-}
-
-/// Walker state that stays valid across peer-hoist discovery passes of
-/// one workspace resolve: the walked tree grows monotonically between
-/// passes ([`crate::WorkspaceTreeCtx::sync_discovery_tree`] rebuilds
-/// the engine's view on the incompatible exceptions), so a subtree
-/// verdict recorded under one importer's walk short-circuits every
-/// compatible revisit — [`Walker::find_hit`] re-validates each
-/// [`PeersCacheItem`] against the current parent context, and a hoisted
-/// provider that newly satisfies a cached missing peer rejects the
-/// stale item. This is the same sharing
-/// [`fn@resolve_peers_workspace`] already applies across importers
-/// within its single final pass.
-#[derive(Debug, Default)]
-pub(crate) struct PeerDiscoveryCaches {
-    node_dep_paths: HashMap<NodeId, DepPath>,
-    pure_pkgs: HashMap<String, DepPath>,
-    peers_cache: HashMap<String, Vec<PeersCacheItem>>,
-    parent_pkgs_of_node: HashMap<NodeId, Arc<HashMap<String, ParentPkgInfo>>>,
-    retained_peer_node_ids: HashSet<NodeId>,
-    peer_provider_children_by_pkg_id: HashMap<String, PeerProviderChildren>,
-    peer_provider_index_peer_names: HashSet<String>,
-}
-
-/// Peer-hoist discovery engine: one persistent tree view + walker
-/// caches shared by every hoist round of a workspace resolve. Replaces
-/// the per-round full snapshot + full [`fn@resolve_peers`] walk that
-/// made multi-importer hoist discovery quadratic in workspace size.
-pub(crate) struct PeerHoistDiscovery {
-    tree: ResolvedTree,
-    caches: PeerDiscoveryCaches,
-    synced_revision: Option<u64>,
-    synced_children_rewrites: Option<u64>,
-}
-
-impl PeerHoistDiscovery {
-    pub(crate) fn new() -> Self {
-        PeerHoistDiscovery {
-            tree: ResolvedTree::default(),
-            caches: PeerDiscoveryCaches::default(),
-            synced_revision: None,
-            synced_children_rewrites: None,
-        }
-    }
-
-    /// Run one discovery pass over `direct` (an importer's current
-    /// direct-dep envelopes), refreshing the persistent tree view from
-    /// `workspace` first when the context changed since the last pass.
-    ///
-    /// A children-ownership handover that rewrote existing occurrence
-    /// nodes ([`crate::WorkspaceTreeCtx::children_rewrites`]) discards
-    /// the whole view: the retained realized children and the walk
-    /// verdicts derived from them predate the rewrite, and an
-    /// incremental sync cannot tell which of them the rewrite
-    /// invalidated.
-    /// `parents_direct` is the importer's full direct set — it defines
-    /// the importer-level peer providers every walked subtree matches
-    /// against. `walk_direct` is the slice actually walked; a hoist
-    /// round that only added new direct deps passes just the additions
-    /// and merges the result into its previous rounds' output.
-    pub(crate) fn discover(
-        &mut self,
-        workspace: &crate::resolve_dependency_tree::WorkspaceTreeCtx,
-        parents_direct: &[DirectDep],
-        walk_direct: &[DirectDep],
-        opts: ResolvePeersOptions,
-    ) -> PeerDiscoveryResult {
-        let revision = workspace.revision();
-        if self.synced_revision != Some(revision) {
-            let children_rewrites = workspace.children_rewrites();
-            let stale =
-                self.synced_children_rewrites.is_some_and(|synced| synced != children_rewrites);
-            if stale || !workspace.sync_discovery_tree(&mut self.tree) {
-                self.tree = ResolvedTree::default();
-                self.caches = PeerDiscoveryCaches::default();
-                let rebuilt = workspace.sync_discovery_tree(&mut self.tree);
-                debug_assert!(rebuilt, "a sync into an empty tree has nothing to conflict with");
-            }
-            self.synced_children_rewrites = Some(children_rewrites);
-            self.synced_revision = Some(revision);
-        }
-        let (result, caches) = discover_peers(
-            &mut self.tree,
-            parents_direct,
-            walk_direct,
-            std::mem::take(&mut self.caches),
-            opts,
-        );
-        self.caches = caches;
-        result
-    }
-}
-
-/// Walk `direct` in discovery mode: peer matching, caches, and issue
-/// collection run exactly as in [`fn@resolve_peers`], but no
-/// [`DependenciesGraph`] is built and no final depPath pass runs. The
-/// per-importer driver mirrors [`fn@resolve_peers_workspace`]'s
-/// per-importer section.
-fn discover_peers(
-    tree: &mut ResolvedTree,
-    parents_direct: &[DirectDep],
-    walk_direct: &[DirectDep],
-    caches: PeerDiscoveryCaches,
-    opts: ResolvePeersOptions,
-) -> (PeerDiscoveryResult, PeerDiscoveryCaches) {
-    let current_provider_sources = vec![CurrentProviderSource {
-        direct_node_ids_by_alias: parents_direct
-            .iter()
-            .map(|dep| (dep.alias.clone(), dep.node_id.clone()))
-            .collect(),
-        declared_direct_dependencies: opts.declared_direct_dependencies.clone(),
-        explicitly_requested_direct_dependencies: opts
-            .explicitly_requested_direct_dependencies
-            .clone(),
-    }];
-    let mut walker =
-        Walker::new(tree, opts, HashMap::default(), current_provider_sources, caches, true);
-
-    let importer_parents = Arc::new(walker.build_importer_parents_from(parents_direct));
-    let parent_chain_names = SharedChain::default();
-    let parent_node_ids = SharedChain::default();
-    let parent_pkg_ids_chain = SharedChain::default();
-    let importer_parent_dep_paths = walker.parent_dep_paths_from_refs(&importer_parents);
-    let (own_direct, provider_direct): (Vec<&DirectDep>, Vec<&DirectDep>) = walk_direct
-        .iter()
-        .partition(|dep| !walker.opts.hoisted_peer_provider_node_ids.contains(&dep.node_id));
-    let mut result = PeerDiscoveryResult::default();
-    for dep in &own_direct {
-        walker.remember_parent_context_if_peer_provider(
-            &dep.alias,
-            &dep.node_id,
-            &importer_parent_dep_paths,
-        );
-    }
-    let mut seen_missing_summaries = HashSet::default();
-    let mut fold_output = |result: &mut PeerDiscoveryResult, output: NodeOutput| {
-        for (peer_alias, peer_node_id) in output.auto_install_resolved_peers {
-            result.resolved_peer_providers_by_alias.insert(peer_alias, peer_node_id);
-        }
-        merge_missing_summary(
-            &mut result.missing_names_by_pkg,
-            &mut seen_missing_summaries,
-            output.subtree_missing_by_pkg,
-        );
-    };
-    for dep in &own_direct {
-        let output = walker.resolve_node(
-            &dep.node_id,
-            &importer_parents,
-            &importer_parent_dep_paths,
-            &parent_chain_names,
-            &parent_node_ids,
-            &parent_pkg_ids_chain,
-        );
-        fold_output(&mut result, output);
-    }
-    // See ResolvePeersOptions::hoisted_peer_provider_node_ids — a
-    // provider is normally resolved at its tree position during the
-    // walk above; only one whose position was pruned still needs the
-    // root-context fallback.
-    for dep in &provider_direct {
-        if walker.visited_this_call.contains(&dep.node_id) {
-            continue;
-        }
-        walker.remember_parent_context_if_peer_provider(
-            &dep.alias,
-            &dep.node_id,
-            &importer_parent_dep_paths,
-        );
-        let output = walker.resolve_node(
-            &dep.node_id,
-            &importer_parents,
-            &importer_parent_dep_paths,
-            &parent_chain_names,
-            &parent_node_ids,
-            &parent_pkg_ids_chain,
-        );
-        fold_output(&mut result, output);
-    }
-    result.peer_dependency_issues = std::mem::take(&mut walker.issues);
-    result.missing_ancestor_pkg_ids = std::mem::take(&mut walker.missing_ancestor_pkg_ids);
-    (result, walker.into_caches())
 }
 
 /// One importer's input to the multi-importer [`fn@resolve_peers_workspace`]
@@ -539,7 +272,7 @@ pub struct WorkspaceResolvePeersResult {
 
 /// Resolve peer dependencies for `tree` and emit a depPath-keyed graph.
 ///
-/// Takes `tree` by `&mut` because lazy [`TreeChildren`] entries are
+/// Takes `tree` by `&mut` because lazy [`crate::TreeChildren`] entries are
 /// realised in-place during the walk — every revisit's `(alias →
 /// NodeId)` children map is allocated on first descent and the
 /// parent's `TreeChildren::Lazy` flips to `Realized` so a second
@@ -568,73 +301,6 @@ pub fn resolve_peers(tree: &mut ResolvedTree, opts: ResolvePeersOptions) -> Reso
         false,
     );
     walker.walk()
-}
-
-pub(crate) fn apply_hoist_missing_scope(
-    result: &mut PeerDiscoveryResult,
-    scope: &HoistMissingScope,
-) {
-    result.peer_dependency_issues.missing.retain(|peer_name, issues| {
-        let ancestor_chains = result.missing_ancestor_pkg_ids.remove(peer_name).unwrap_or_default();
-        *issues = std::mem::take(issues)
-            .into_iter()
-            .zip(ancestor_chains)
-            .filter_map(|(issue, ancestor_pkg_ids)| {
-                (!scope.suppresses_iter(ancestor_pkg_ids.iter(), peer_name)).then_some(issue)
-            })
-            .collect();
-        !issues.is_empty()
-    });
-}
-
-/// The current-provider sources visible while walking `importer`: its
-/// own direct dependencies, plus the workspace root importer's when
-/// that is a different project. The workspace path carries no
-/// declared/explicitly-requested alias sets yet, so those guards stay
-/// empty here.
-fn importer_provider_sources(
-    importer: &ImporterPeerInput,
-    root_importer: Option<&ImporterPeerInput>,
-) -> Vec<CurrentProviderSource> {
-    let source_of = |importer: &ImporterPeerInput| CurrentProviderSource {
-        direct_node_ids_by_alias: importer
-            .direct
-            .iter()
-            .map(|dep| (dep.alias.clone(), dep.node_id.clone()))
-            .collect(),
-        declared_direct_dependencies: HashSet::default(),
-        explicitly_requested_direct_dependencies: HashSet::default(),
-    };
-    let mut sources = vec![source_of(importer)];
-    if let Some(root) = root_importer
-        && root.id != importer.id
-    {
-        sources.push(source_of(root));
-    }
-    sources
-}
-
-/// The upstream `getNodeIdsByPreviousDepPath`: first node (in
-/// deterministic id order) claiming each `previous_dep_path`. Empty
-/// unless a locked-peer reuse pass was requested.
-fn build_node_ids_by_previous_dep_path(
-    tree: &ResolvedTree,
-    opts: &ResolvePeersOptions,
-) -> HashMap<DepPath, NodeId> {
-    let mut map = HashMap::default();
-    if opts.resolved_peer_provider_paths.is_none() {
-        return map;
-    }
-    let mut node_ids: Vec<&NodeId> = tree.dependencies_tree.keys().collect();
-    node_ids.sort_by_key(|node_id| node_id_sort_key(node_id));
-    for node_id in node_ids {
-        if let Some(previous) = tree.dependencies_tree[node_id].previous_dep_path.as_ref()
-            && !map.contains_key(previous)
-        {
-            map.insert(previous.clone(), node_id.clone());
-        }
-    }
-    map
 }
 
 /// Resolve peer dependencies for every importer in `importers` against
@@ -795,3140 +461,57 @@ pub fn resolve_peers_workspace(
     }
 }
 
-/// Per-name entry in the propagating [`ParentRefs`] map.
-#[derive(Debug, Clone)]
-struct ParentRef {
-    version: String,
-    /// `None` for top-level deps that were already installed. Pacquet
-    /// doesn't surface those yet — `None` only appears on the
-    /// importer-level cycle-break fallback, where the `name@version`
-    /// form of the peer-id is the only useful representation.
-    node_id: Option<NodeId>,
-    /// Local install name in `node_modules`. May differ from the
-    /// package's real name for npm-alias entries.
-    alias: Option<String>,
-    /// Depth at which this parent was added. Threaded into
-    /// [`ParentPkgInfo`] so [`Walker::parent_packages_match`] can
-    /// apply the depth-equality fallback when peer dependencies are
-    /// shadowed across occurrences.
-    depth: i32,
-    /// Per-name shadowing counter. Incremented when a same-name
-    /// parent is added at a deeper walk that doesn't match the
-    /// existing entry. Used by [`Walker::parent_packages_match`] to
-    /// detect shadowed peers.
-    occurrence: u32,
-}
-
-/// `name → ParentRef` map propagated down the walk. Entries are indexed
-/// by both the package's real name and its alias when the two differ —
-/// `react-dom@npm:next` resolves a `peerDependencies.react-dom`
-/// requirement against the alias and `peerDependencies.next` against
-/// the real name.
-type ParentRefs = HashMap<String, ParentRef>;
-
-enum FastProvider<'a> {
-    Missing,
-    Inherited(&'a ParentRef),
-    Child(&'a str),
-    Ambiguous,
-}
-
-enum FastCacheMatch {
-    Match,
-    NoMatch,
-    Ambiguous,
-}
-
-#[derive(Debug, Clone)]
-struct SharedChain<Element>(Option<Arc<SharedChainLink<Element>>>);
-
-#[derive(Debug)]
-struct SharedChainLink<Element> {
-    value: Element,
-    parent: Option<Arc<SharedChainLink<Element>>>,
-}
-
-impl<Element> Default for SharedChain<Element> {
-    fn default() -> Self {
-        SharedChain(None)
-    }
-}
-
-impl<Element> SharedChain<Element> {
-    fn pushed(&self, value: Element) -> Self {
-        SharedChain(Some(Arc::new(SharedChainLink { value, parent: self.0.clone() })))
-    }
-
-    fn iter(&self) -> SharedChainIter<'_, Element> {
-        SharedChainIter { next: self.0.as_deref() }
-    }
-}
-
-impl<Element: PartialEq> SharedChain<Element> {
-    fn contains(&self, value: &Element) -> bool {
-        self.iter().any(|item| item == value)
-    }
-}
-
-impl<Element: Clone> SharedChain<Element> {
-    fn to_root_vec(&self) -> Vec<Element> {
-        let mut values: Vec<Element> = self.iter().cloned().collect();
-        values.reverse();
-        values
-    }
-}
-
-struct SharedChainIter<'a, Element> {
-    next: Option<&'a SharedChainLink<Element>>,
-}
-
-impl<'a, Element> Iterator for SharedChainIter<'a, Element> {
-    type Item = &'a Element;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let link = self.next?;
-        self.next = link.parent.as_deref();
-        Some(&link.value)
-    }
-}
-
-fn scoped_hoisted_optional_parent_refs(
-    parent_refs: &ParentRefs,
-    locked_peer_names: &HashSet<String>,
-    hoisted_optional_peer_node_ids: &HashSet<NodeId>,
-) -> ParentRefs {
-    parent_refs
-        .iter()
-        .filter(|(name, parent)| {
-            parent.node_id.as_ref().is_none_or(|parent_node_id| {
-                !hoisted_optional_peer_node_ids.contains(parent_node_id)
-                    || locked_peer_names.contains(*name)
-            })
-        })
-        .map(|(name, parent)| (name.clone(), parent.clone()))
-        .collect()
-}
-
-/// One importer whose direct dependencies count as "current" peer
-/// providers for the must-win guard of locked-peer-provider reuse.
-struct CurrentProviderSource {
-    direct_node_ids_by_alias: HashMap<String, NodeId>,
-    declared_direct_dependencies: HashSet<String>,
-    explicitly_requested_direct_dependencies: HashSet<String>,
-}
-
-#[derive(Debug, Default)]
-struct PeerProviderChildren {
-    relevant_edge_indices: Vec<usize>,
-    edge_indices_by_name: HashMap<String, Vec<usize>>,
-}
-
-#[derive(Clone, Copy)]
-struct FinalPeerContext<'a> {
-    scc_of: &'a HashMap<NodeId, usize>,
-    cyclic_peer_names: &'a HashSet<String>,
-}
-
-struct UndoRealize {
-    newly_inserted: Vec<NodeId>,
-    prev_parent_ids: AncestorIds,
-}
-
-/// Combines preview and final-materialization undo logs for the same node.
-/// Both logs restore the same pre-realization ancestor chain; previewing
-/// does not change the node's lazy parent state.
-fn merge_realize_undo(
-    first: Option<UndoRealize>,
-    second: Option<UndoRealize>,
-) -> Option<UndoRealize> {
-    match (first, second) {
-        (None, undo) | (undo, None) => undo,
-        (Some(mut first), Some(second)) => {
-            first.newly_inserted.extend(second.newly_inserted);
-            Some(first)
-        }
-    }
-}
-
-struct Walker<'tree> {
-    tree: &'tree mut ResolvedTree,
-    opts: ResolvePeersOptions,
-    graph: DependenciesGraph,
-    issues: PeerDependencyIssues,
-    missing_ancestor_pkg_ids: HashMap<String, Vec<SharedChain<String>>>,
-    /// `NodeId → DepPath` once a node has been walked. Lets repeated
-    /// visits (an importer-direct dep that's also reached transitively)
-    /// reuse the already-computed depPath.
-    node_dep_paths: HashMap<NodeId, DepPath>,
-    /// Peers each node and its subtree resolved against ancestors —
-    /// the "unknown resolved peers" propagated up so a parent can fold
-    /// its descendants' peer dependencies into its own peer suffix.
-    /// Indexed by `NodeId`; value's keys are peer aliases.
-    node_external_peers: HashMap<NodeId, Arc<HashMap<String, NodeId>>>,
-    /// Peers each node and its subtree declared but couldn't find.
-    /// Indexed by `NodeId`; value's keys are peer aliases.
-    node_missing_peers: HashMap<NodeId, Arc<HashMap<String, MissingPeerInfo>>>,
-    /// Peers each node's children declared but couldn't find.
-    /// Indexed by `NodeId`; value's keys are peer aliases.
-    node_missing_peers_of_children: HashMap<NodeId, Arc<HashMap<String, MissingPeerInfo>>>,
-    /// Resolver-stage real peer providers seen while walking the tree.
-    /// This intentionally excludes `peerDependenciesMeta`-only entries:
-    /// the auto-install pass reads real `peerDependencies` entries only.
-    resolved_peer_providers_by_alias: BTreeMap<String, NodeId>,
-    /// Stack of nodes currently being walked. Re-entry on a node here
-    /// is a cycle — the recursion bottoms out with a `name@version`
-    /// peer-id and the original visit drives the actual graph insert.
-    in_progress: HashSet<NodeId>,
-    /// Graph edges whose target `NodeId` had no `DepPath` yet at the
-    /// time we built the parent's `graph_children` map — typically
-    /// because the target is a later sibling direct dep that the walker
-    /// hasn't reached yet. `walk()` drains this list once every direct
-    /// dep is walked and patches the recorded entries with the now-known
-    /// `DepPath`. Without this post-pass the install layer
-    /// would walk the parent's `children` map and find no symlink edge
-    /// for the child, leaving the package without it in its slot.
-    pending_peer_edges: Vec<PendingPeerEdge>,
-    /// Set of `pkgIdWithPatchHash` values whose full subtree resolved
-    /// with zero external peers and zero missing peers. A revisit of
-    /// any such package whose own `peerDependencies` is empty
-    /// short-circuits with `depPath = pkgIdWithPatchHash` — no
-    /// recursion, no peersCache lookup.
-    /// Populated bottom-up: a node is added when its local `is_pure`
-    /// flag is true after its own walk completes.
-    pure_pkgs: HashMap<String, DepPath>,
-    /// Per-`pkgIdWithPatchHash` cached results from earlier walks of
-    /// non-pure subtrees. Each cache item records the `depPath`, the
-    /// external `(peer_name → NodeId)` map, and the `(peer_name →
-    /// info)` missing set produced by one specific parent peer
-    /// context. [`Walker::find_hit`] iterates the bucket and accepts
-    /// the first item whose cached context is compatible with the
-    /// current call's `child_parent_refs`.
-    ///
-    /// [`Walker::find_hit`] calls [`Walker::parent_packages_match`] for
-    /// the deep check, which compares against the parent context
-    /// snapshots held in [`Walker::parent_pkgs_of_node`].
-    peers_cache: HashMap<String, Vec<PeersCacheItem>>,
-    /// Per-`NodeId` snapshot of the parent peer context (peer-relevant
-    /// names → [`ParentPkgInfo`]) recorded at the moment the walker
-    /// first descended into that node. Backs
-    /// [`Walker::parent_packages_match`]: a [`PeersCacheItem`] is a
-    /// cache hit only when each of its resolved-peer `NodeId`s has an
-    /// entry here whose recorded parent context still matches the
-    /// current walk's `parent_refs` (or, for `purePkgs` peers, the
-    /// presence-and-pkg-id match short-circuit).
-    parent_pkgs_of_node: HashMap<NodeId, Arc<HashMap<String, ParentPkgInfo>>>,
-    retained_peer_node_ids: HashSet<NodeId>,
-    /// Per-`NodeId` snapshot captured at graph-insert time, consumed by
-    /// the post-walk [`Walker::build_final_dep_paths`] /
-    /// [`Walker::build_final_graph`] pass. See [`NodeRecord`].
-    node_records: HashMap<NodeId, NodeRecord>,
-    next_record_order: u64,
-    /// Reverse index over the tree nodes' `previous_dep_path`, built
-    /// only when [`ResolvePeersOptions::resolved_peer_provider_paths`]
-    /// is set. The upstream `nodeIdsByPreviousDepPath`.
-    node_ids_by_previous_dep_path: HashMap<DepPath, NodeId>,
-    /// Importers whose direct dependencies count as "current" peer
-    /// providers for the must-win guard. Swapped per importer by the
-    /// workspace entry point.
-    current_provider_sources: Vec<CurrentProviderSource>,
-    /// `true` for a peer-hoist discovery pass: the walk records no
-    /// graph entries, node records, or pending edges, and the caller
-    /// runs none of the final depPath/graph passes. Everything that
-    /// decides *what* resolves or goes missing is unchanged.
-    discovery: bool,
-    /// Nodes this call resolved (any return path except the cycle
-    /// re-entry). Distinguishes them from nodes only known through the
-    /// persistent [`PeerDiscoveryCaches`], so the pruned-provider
-    /// fallback keeps its per-call meaning.
-    visited_this_call: HashSet<NodeId>,
-    packages_by_id: HashMap<String, Arc<ResolvedPackage>>,
-    peer_provider_children_by_pkg_id: HashMap<String, PeerProviderChildren>,
-    peer_provider_index_peer_names: HashSet<String>,
-    empty_resolved_peers: Arc<HashMap<String, NodeId>>,
-    empty_missing_peers: Arc<HashMap<String, MissingPeerInfo>>,
-}
-
-impl<'tree> Walker<'tree> {
-    fn new(
-        tree: &'tree mut ResolvedTree,
-        opts: ResolvePeersOptions,
-        node_ids_by_previous_dep_path: HashMap<DepPath, NodeId>,
-        current_provider_sources: Vec<CurrentProviderSource>,
-        caches: PeerDiscoveryCaches,
-        discovery: bool,
-    ) -> Self {
-        let PeerDiscoveryCaches {
-            node_dep_paths,
-            pure_pkgs,
-            peers_cache,
-            parent_pkgs_of_node,
-            retained_peer_node_ids,
-            mut peer_provider_children_by_pkg_id,
-            mut peer_provider_index_peer_names,
-        } = caches;
-        if peer_provider_index_peer_names != tree.all_peer_dep_names {
-            peer_provider_children_by_pkg_id.clear();
-            peer_provider_index_peer_names.clone_from(&tree.all_peer_dep_names);
-        }
-        for (pkg_id, children) in &tree.children_by_id {
-            if peer_provider_children_by_pkg_id.contains_key(pkg_id) {
-                continue;
-            }
-            let mut providers = PeerProviderChildren::default();
-            for (edge_index, edge) in children.iter().enumerate() {
-                let Some(pkg) = tree.packages.get(&edge.pkg_id) else { continue };
-                let real_name = pkg_name_version(&pkg.result).0;
-                let alias_is_peer = tree.all_peer_dep_names.contains(&edge.alias);
-                let real_name_is_peer = tree.all_peer_dep_names.contains(&real_name);
-                if !alias_is_peer && !real_name_is_peer {
-                    continue;
-                }
-                providers.relevant_edge_indices.push(edge_index);
-                if alias_is_peer {
-                    providers
-                        .edge_indices_by_name
-                        .entry(edge.alias.clone())
-                        .or_default()
-                        .push(edge_index);
-                }
-                if real_name_is_peer && real_name != edge.alias {
-                    providers.edge_indices_by_name.entry(real_name).or_default().push(edge_index);
-                }
-            }
-            peer_provider_children_by_pkg_id.insert(pkg_id.clone(), providers);
-        }
-        Walker {
-            tree,
-            opts,
-            graph: DependenciesGraph::default(),
-            issues: PeerDependencyIssues::default(),
-            missing_ancestor_pkg_ids: HashMap::default(),
-            node_dep_paths,
-            node_external_peers: HashMap::default(),
-            node_missing_peers: HashMap::default(),
-            node_missing_peers_of_children: HashMap::default(),
-            resolved_peer_providers_by_alias: BTreeMap::new(),
-            in_progress: HashSet::default(),
-            pending_peer_edges: Vec::new(),
-            pure_pkgs,
-            peers_cache,
-            parent_pkgs_of_node,
-            retained_peer_node_ids,
-            node_records: HashMap::default(),
-            next_record_order: 0,
-            node_ids_by_previous_dep_path,
-            current_provider_sources,
-            discovery,
-            visited_this_call: HashSet::default(),
-            packages_by_id: HashMap::default(),
-            peer_provider_children_by_pkg_id,
-            peer_provider_index_peer_names,
-            empty_resolved_peers: Arc::new(HashMap::default()),
-            empty_missing_peers: Arc::new(HashMap::default()),
-        }
-    }
-
-    fn into_caches(self) -> PeerDiscoveryCaches {
-        PeerDiscoveryCaches {
-            node_dep_paths: self.node_dep_paths,
-            pure_pkgs: self.pure_pkgs,
-            peers_cache: self.peers_cache,
-            parent_pkgs_of_node: self.parent_pkgs_of_node,
-            retained_peer_node_ids: self.retained_peer_node_ids,
-            peer_provider_children_by_pkg_id: self.peer_provider_children_by_pkg_id,
-            peer_provider_index_peer_names: self.peer_provider_index_peer_names,
-        }
-    }
-}
-
-/// Per-peer-name snapshot stored on [`Walker::parent_pkgs_of_node`].
-///
-/// `pkg_id` is `None` for parents that came in without a real
-/// `NodeId` (the importer-level `topParents` path); those
-/// fall back to a pure `version` comparison.
-#[derive(Debug, Clone)]
-struct ParentPkgInfo {
-    pkg_id: Option<String>,
-    version: Option<String>,
-    depth: i32,
-    occurrence: u32,
-}
-
-/// One cached resolution of a non-pure subtree: the part of a walk's
-/// verdict that holds in any compatible parent context, so a revisit —
-/// by the same walk or another importer's — can reuse it.
-///
-/// `dep_path` is the value [`Walker::resolve_node`] would otherwise
-/// recompute. `resolved_peers` is the external peer set (excluding
-/// peers satisfied by this node's own children) — [`Walker::find_hit`]
-/// uses it as the cache-match key against the current parent context.
-/// `missing_peers` is the set of unmet peer requirements the original
-/// walk surfaced — when a cache item carries a missing peer that the
-/// current parent context *does* provide, the contexts are
-/// incompatible and the item must be rejected. `missing_peers_of_children`
-/// is the subset exposed as the package's children report.
-///
-/// The peer *providers* the walk resolved to are deliberately not part
-/// of the verdict: they belong to the resolving walk's own context —
-/// pnpm's resolver gives a not-new package `resolvedPeers: {}`
-/// (`resolveDependencies.ts`) — so only the walk that first resolved a
-/// subtree promotes its providers to importer level.
-#[derive(Debug)]
-struct PeersCacheItem {
-    dep_path: DepPath,
-    resolved_peers: Arc<HashMap<String, NodeId>>,
-    missing_peers: Arc<HashMap<String, MissingPeerInfo>>,
-    missing_peers_of_children: Arc<HashMap<String, MissingPeerInfo>>,
-    /// See [`NodeOutput::subtree_missing_by_pkg`]. Replayed on a cache
-    /// hit so a discovery pass that never descends into the cached
-    /// subtree still reports the same per-package missing breakdown a
-    /// full walk of it would.
-    subtree_missing_by_pkg: SubtreeMissingByPkg,
-}
-
-struct CachedNodeOutput {
-    output: NodeOutput,
-    missing_peers_of_children: Arc<HashMap<String, MissingPeerInfo>>,
-}
-
-enum DeferredChildResolution {
-    Pure(DepPath),
-    Cached(CachedNodeOutput),
-    Materialize(String),
-}
-
-struct CacheHitContext<'a> {
-    node_id: &'a NodeId,
-    tree_node_depth: i32,
-    parent_chain_names: &'a SharedChain<String>,
-    parent_pkg_ids_chain: &'a SharedChain<String>,
-    preview_undo: Option<UndoRealize>,
-}
-
-struct DeferredChildContext<'a> {
-    edge: &'a ChildEdge,
-    node_id: NodeId,
-    parent_ids: &'a AncestorIds,
-    parent_refs: &'a Arc<ParentRefs>,
-    parent_dep_paths: &'a Arc<HashMap<String, ParentPkgInfo>>,
-    chain_names: &'a SharedChain<String>,
-    parent_node_ids: &'a SharedChain<NodeId>,
-    parent_pkg_ids: &'a SharedChain<String>,
-    depth: i32,
-}
-
-impl PeersCacheItem {
-    /// The [`NodeOutput`] a cache hit hands the reusing walk; output
-    /// fields the verdict doesn't carry come back empty.
-    fn to_node_output(&self) -> NodeOutput {
-        NodeOutput {
-            dep_path: self.dep_path.clone(),
-            external_resolved_peers: Arc::clone(&self.resolved_peers),
-            auto_install_resolved_peers: HashMap::default(),
-            missing_peers: Arc::clone(&self.missing_peers),
-            subtree_missing_by_pkg: self.subtree_missing_by_pkg.clone(),
-        }
-    }
-
-    fn to_cached_node_output(&self) -> CachedNodeOutput {
-        CachedNodeOutput {
-            output: self.to_node_output(),
-            missing_peers_of_children: Arc::clone(&self.missing_peers_of_children),
-        }
-    }
-}
-
-/// One `parent → child` edge whose target wasn't walked yet at the
-/// time the parent's `graph_children` was built. Patched up by
-/// [`Walker::patch_pending_peer_edges`] after the main walk completes.
-struct PendingPeerEdge {
-    parent_dep_path: DepPath,
-    child_alias: String,
-    child_node_id: NodeId,
-}
-
-/// Per-`NodeId` data captured during the walk so the post-walk
-/// [`Walker::build_final_dep_paths`] pass can recompute each node's
-/// depPath with its resolved peers' *full* suffixes, collapsing to
-/// `name@version` only for genuinely detected cycles.
-///
-/// The walk itself is left untouched: it still computes the provisional
-/// depPaths that [`Walker::find_hit`] reads, so peer-resolution and
-/// cache decisions are byte-for-byte identical. Only the rendered
-/// depPaths change, which is why a node whose suffix was previously
-/// collapsed by the cycle fallback now splits into its own graph entry.
-struct NodeRecord {
-    /// `alias → child/peer NodeId` edges, in the same shape the inline
-    /// `graph_children` map carries (children overlaid with resolved-peer
-    /// edges) but holding `NodeIds`, so the rebuild can map each edge to
-    /// its final depPath.
-    edges: BTreeMap<String, NodeId>,
-    optional_child_aliases: HashSet<String>,
-    transitive_peer_dependencies: HashSet<String>,
-    depth: i32,
-    installable: bool,
-    is_pure: bool,
-    order: u64,
-}
-
-/// Sentinel for "this node's subtree is still missing peer `X`". The
-/// `range` + `optional` payload is recorded for a future `peersCache`
-/// lookup, but the issue-collection path uses
-/// [`PeerDependencyIssues::missing`] directly, so neither field is read
-/// after construction yet.
-#[derive(Debug, Clone)]
-struct MissingPeerInfo {
-    #[allow(dead_code, reason = "future peersCache validation")]
-    range: String,
-    #[allow(dead_code, reason = "future peersCache validation")]
-    optional: bool,
-}
-
-/// Persistent summary of missing peers in a subtree. Child summaries
-/// are shared with their parents instead of repeatedly copying every
-/// descendant's package map into each ancestor and cache entry.
-#[derive(Debug)]
-struct MissingSummary {
-    own: Option<(String, HashSet<String>)>,
-    children: Vec<Arc<MissingSummary>>,
-}
-
-type SubtreeMissingByPkg = Option<Arc<MissingSummary>>;
-
-/// Output of [`Walker::resolve_node`] — the per-node result the parent
-/// folds into its own state.
-struct NodeOutput {
-    dep_path: DepPath,
-    /// Peers that this node + its subtree resolved against ancestors.
-    /// Excludes peers resolved against this node's own children (those
-    /// are absorbed into the children's depPaths).
-    external_resolved_peers: Arc<HashMap<String, NodeId>>,
-    /// Real `peerDependencies` resolved anywhere in this node's
-    /// subtree. This feeds the auto-install-peers loop.
-    auto_install_resolved_peers: HashMap<String, NodeId>,
-    missing_peers: Arc<HashMap<String, MissingPeerInfo>>,
-    /// [`ResolvePeersResult::missing_names_by_pkg`]'s per-subtree
-    /// slice, propagated bottom-up so discovery can aggregate it from
-    /// the importer's direct deps alone.
-    subtree_missing_by_pkg: SubtreeMissingByPkg,
-}
-
-enum ChildAliases<'a> {
-    Realized(&'a BTreeMap<String, NodeId>),
-    Deferred(&'a [ChildEdge]),
-}
-
-impl ChildAliases<'_> {
-    fn contains(&self, alias: &str) -> bool {
-        match self {
-            ChildAliases::Realized(children) => children.contains_key(alias),
-            ChildAliases::Deferred(children) => children.iter().any(|edge| edge.alias == alias),
-        }
-    }
-}
-
-#[derive(Default)]
-struct ChildOutputs {
-    external_peers: HashMap<String, NodeId>,
-    auto_install_resolved_peers: HashMap<String, NodeId>,
-    missing_peers: HashMap<String, MissingPeerInfo>,
-    dep_paths: BTreeMap<String, DepPath>,
-    missing_summaries: Vec<Arc<MissingSummary>>,
-}
-
-impl ChildOutputs {
-    fn push(
-        &mut self,
-        alias: &str,
-        output: NodeOutput,
-        child_aliases: &ChildAliases<'_>,
-        collect_dep_paths: bool,
-    ) {
-        let NodeOutput {
-            dep_path,
-            external_resolved_peers,
-            auto_install_resolved_peers,
-            missing_peers,
-            subtree_missing_by_pkg,
-        } = output;
-        if let Some(summary) = subtree_missing_by_pkg
-            && !self.missing_summaries.iter().any(|existing| Arc::ptr_eq(existing, &summary))
-        {
-            self.missing_summaries.push(summary);
-        }
-        if collect_dep_paths {
-            self.dep_paths.insert(alias.to_string(), dep_path);
-        }
-        self.auto_install_resolved_peers.extend(auto_install_resolved_peers);
-        for (peer_alias, peer_node_id) in external_resolved_peers.iter() {
-            if !child_aliases.contains(peer_alias) {
-                self.external_peers.insert(peer_alias.clone(), peer_node_id.clone());
-            }
-        }
-        self.missing_peers
-            .extend(missing_peers.iter().map(|(name, info)| (name.clone(), info.clone())));
-    }
-}
-
-impl Walker<'_> {
-    fn walk(mut self) -> ResolvePeersResult {
-        let importer_parents = Arc::new(self.build_importer_parents());
-        let parent_chain_names = SharedChain::default();
-        let parent_node_ids = SharedChain::default();
-        let parent_pkg_ids_chain = SharedChain::default();
-        let mut direct_by_alias = BTreeMap::new();
-        // Clone direct deps into an owned `Vec` so the recursion
-        // below can mutate `self.tree` (realising lazy children)
-        // without conflicting with this loop's borrow of
-        // `self.tree.direct`.
-        let direct: Vec<DirectDep> = self.tree.direct.clone();
-        let importer_parent_dep_paths = self.parent_dep_paths_from_refs(&importer_parents);
-        let (own_direct, provider_direct): (Vec<&DirectDep>, Vec<&DirectDep>) = direct
+/// The current-provider sources visible while walking `importer`: its
+/// own direct dependencies, plus the workspace root importer's when
+/// that is a different project. The workspace path carries no
+/// declared/explicitly-requested alias sets yet, so those guards stay
+/// empty here.
+fn importer_provider_sources(
+    importer: &ImporterPeerInput,
+    root_importer: Option<&ImporterPeerInput>,
+) -> Vec<CurrentProviderSource> {
+    let source_of = |importer: &ImporterPeerInput| CurrentProviderSource {
+        direct_node_ids_by_alias: importer
+            .direct
             .iter()
-            .partition(|dep| !self.opts.hoisted_peer_provider_node_ids.contains(&dep.node_id));
-        for dep in &own_direct {
-            self.remember_parent_context_if_peer_provider(
-                &dep.alias,
-                &dep.node_id,
-                &importer_parent_dep_paths,
-            );
-        }
-        for dep in &own_direct {
-            let output = self.resolve_node(
-                &dep.node_id,
-                &importer_parents,
-                &importer_parent_dep_paths,
-                &parent_chain_names,
-                &parent_node_ids,
-                &parent_pkg_ids_chain,
-            );
-            for (peer_alias, peer_node_id) in output.auto_install_resolved_peers {
-                self.resolved_peer_providers_by_alias.insert(peer_alias, peer_node_id);
-            }
-        }
-        // See ResolvePeersOptions::hoisted_peer_provider_node_ids — a
-        // provider is normally resolved at its tree position during the
-        // walk above; only one whose position was pruned still needs the
-        // root-context fallback.
-        for dep in &provider_direct {
-            if self.visited_this_call.contains(&dep.node_id) {
-                continue;
-            }
-            self.remember_parent_context_if_peer_provider(
-                &dep.alias,
-                &dep.node_id,
-                &importer_parent_dep_paths,
-            );
-            let output = self.resolve_node(
-                &dep.node_id,
-                &importer_parents,
-                &importer_parent_dep_paths,
-                &parent_chain_names,
-                &parent_node_ids,
-                &parent_pkg_ids_chain,
-            );
-            for (peer_alias, peer_node_id) in output.auto_install_resolved_peers {
-                self.resolved_peer_providers_by_alias.insert(peer_alias, peer_node_id);
-            }
-        }
-        self.patch_pending_peer_edges();
-        // Recompute depPaths so each resolved peer carries its full
-        // suffix (the cycle fallback during the walk collapses peers
-        // that are walk-ancestors), then rebuild the graph from the
-        // per-node records keyed by the corrected depPaths.
-        let final_dep_paths = self.build_final_dep_paths();
-        for dep in &direct {
-            let dep_path = self.final_dep_path_of(&dep.node_id, &final_dep_paths);
-            let dep_path = importer_relative_link_dep_path(
-                &dep_path,
-                self.opts.lockfile_dir.as_deref(),
-                self.opts.project_dir.as_deref(),
-            );
-            direct_by_alias.insert(dep.alias.clone(), dep_path);
-        }
-        let graph = self.build_final_graph(&final_dep_paths);
-        let paths_by_node_id = self.final_paths_by_node_id(&final_dep_paths);
-        let resolved_peer_providers_by_alias = self.resolved_peer_providers_by_alias;
-        let mut missing_names_by_pkg: HashMap<String, HashSet<String>> = HashMap::default();
-        for (node_id, missing) in &self.node_missing_peers_of_children {
-            let Some(tree_node) = self.tree.dependencies_tree.get(node_id) else { continue };
-            missing_names_by_pkg
-                .entry(tree_node.resolved_package_id.clone())
-                .or_default()
-                .extend(missing.keys().cloned());
-        }
-        ResolvePeersResult {
-            graph,
-            direct_dependencies_by_alias: direct_by_alias,
-            resolved_peer_providers_by_alias,
-            peer_dependency_issues: self.issues,
-            missing_names_by_pkg,
-            paths_by_node_id,
-        }
-    }
-
-    /// The upstream `pathsByNodeId`: every walked node's final
-    /// `DepPath`. Empty unless
-    /// [`ResolvePeersOptions::collect_paths_by_node_id`] asked for it.
-    fn final_paths_by_node_id(
-        &self,
-        final_dep_paths: &HashMap<NodeId, DepPath>,
-    ) -> HashMap<NodeId, DepPath> {
-        if !self.opts.collect_paths_by_node_id {
-            return HashMap::default();
-        }
-        self.node_dep_paths
-            .keys()
-            .map(|node_id| (node_id.clone(), self.final_dep_path_of(node_id, final_dep_paths)))
-            .collect()
-    }
-
-    /// Fill in `graph_children` edges that were skipped during the main
-    /// walk because the peer target's `DepPath` hadn't been computed
-    /// yet. Each direct dep's subtree is fully walked by the time
-    /// `walk()` drains this list, so every peer that was reachable
-    /// from an ancestor's [`ParentRefs`] has a `DepPath` now. Peers that
-    /// still don't resolve here came from a `parent_chain` outside the
-    /// walked set — there's nothing to patch, and the absence already
-    /// surfaced via [`PeerDependencyIssues::missing`].
-    fn patch_pending_peer_edges(&mut self) {
-        for edge in std::mem::take(&mut self.pending_peer_edges) {
-            let Some(child_dep_path) = self.node_dep_paths.get(&edge.child_node_id).cloned() else {
-                continue;
-            };
-            if let Some(node) = self.graph.get_mut(&edge.parent_dep_path) {
-                // `entry().or_insert` rather than unconditional insert:
-                // if a later walk of the same `dep_path` already
-                // populated the edge (e.g. via the cycle path), we
-                // don't want to overwrite a more specific entry.
-                node.children.entry(edge.child_alias).or_insert(child_dep_path);
-            }
-        }
-    }
-
-    /// Build the seed [`ParentRefs`] from the importer's direct deps so
-    /// a direct dep's peer requirements can be satisfied by a sibling
-    /// direct dep.
-    ///
-    /// `link:` direct deps whose target lives outside
-    /// [`ResolvePeersOptions::lockfile_dir`] are seeded with a node id
-    /// rewritten to `link:<rel-from-lockfile_dir-to-modules_dir>/<alias>`
-    /// when [`ResolvePeersOptions::exclude_links_from_lockfile`] is on
-    /// — keeping the peer-suffix segment stable across machines
-    /// regardless of the absolute path of the external link.
-    fn build_importer_parents(&self) -> ParentRefs {
-        self.build_importer_parents_from(&self.tree.direct)
-    }
-
-    /// Whether a dependency installed under `alias` can provide a peer:
-    /// its alias or its real package name (the two differ for npm-alias
-    /// deps like `peer-c1@npm:@pnpm.e2e/peer-c@2.0.0`) is declared as a
-    /// peer somewhere in the tree.
-    fn is_peer_relevant(&self, alias: &str, pkg: &ResolvedPackage) -> bool {
-        if self.tree.all_peer_dep_names.contains(alias) {
-            return true;
-        }
-        let (real_name, _) = pkg_name_version(&pkg.result);
-        self.tree.all_peer_dep_names.contains(&real_name)
-    }
-
-    /// Same as [`Self::build_importer_parents`] but seeds from an
-    /// externally-supplied direct-deps slice — used by the multi-importer
-    /// [`fn@resolve_peers_workspace`] where each importer's `direct`
-    /// lives outside [`ResolvedTree`].
-    fn build_importer_parents_from(&self, direct_deps: &[DirectDep]) -> ParentRefs {
-        let mut refs = ParentRefs::default();
-        for direct in direct_deps {
-            let Some(tree_node) = self.tree.dependencies_tree.get(&direct.node_id) else {
-                continue;
-            };
-            let Some(pkg) = self.tree.packages.get(&tree_node.resolved_package_id) else {
-                continue;
-            };
-            if !self.is_peer_relevant(&direct.alias, pkg) {
-                continue;
-            }
-            let parent_node_id = remap_link_node_id(&self.opts, &direct.alias, &pkg.result)
-                .unwrap_or_else(|| direct.node_id.clone());
-            insert_parent_ref(&mut refs, &direct.alias, parent_node_id, pkg, tree_node.depth);
-        }
-        refs
-    }
-
-    fn resolve_node(
-        &mut self,
-        node_id: &NodeId,
-        parent_parent_refs: &Arc<ParentRefs>,
-        parent_dep_paths: &Arc<HashMap<String, ParentPkgInfo>>,
-        parent_chain_names: &SharedChain<String>,
-        parent_node_ids: &SharedChain<NodeId>,
-        parent_pkg_ids_chain: &SharedChain<String>,
-    ) -> NodeOutput {
-        // `purePkgs` fast-path. When the subtree below this
-        // `pkgIdWithPatchHash` resolved with zero external peers and
-        // zero missing peers on a previous walk, AND this package
-        // itself declares no `peerDependencies`, the `depPath` is the
-        // bare `pkgIdWithPatchHash` regardless of parent context.
-        // Skip recursion entirely.
-        //
-        let context_free_dep_path =
-            self.tree.dependencies_tree.get(node_id).and_then(|tree_node| {
-                if tree_node.depth == -1 {
-                    return Some((
-                        tree_node.depth,
-                        DepPath::from(tree_node.resolved_package_id.clone()),
-                    ));
-                }
-                let dep_path = self.pure_pkgs.get(&tree_node.resolved_package_id)?;
-                if !self.tree.packages[&tree_node.resolved_package_id].peer_dependencies.is_empty()
-                    || (!self.discovery
-                        && self
-                            .graph
-                            .get(dep_path)
-                            .is_none_or(|graph_node| graph_node.depth > tree_node.depth))
-                {
-                    return None;
-                }
-                Some((tree_node.depth, dep_path.clone()))
-            });
-        if let Some((tree_node_depth, dep_path)) = context_free_dep_path {
-            self.remember_resolved_node(node_id, &dep_path);
-            if let Some(node) = self.graph.get_mut(&dep_path)
-                && node.depth > tree_node_depth
-            {
-                node.depth = tree_node_depth;
-            }
-            return NodeOutput {
-                dep_path,
-                external_resolved_peers: Arc::clone(&self.empty_resolved_peers),
-                auto_install_resolved_peers: HashMap::default(),
-                missing_peers: Arc::clone(&self.empty_missing_peers),
-                subtree_missing_by_pkg: None,
-            };
-        }
-
-        if self.in_progress.contains(node_id) {
-            // Cycle: bottom out with the bare `pkgIdWithPatchHash` as
-            // the depPath. The original visit (still on the stack) will
-            // compute the real depPath and insert it into
-            // `node_dep_paths`. Returning the bare id here ensures the
-            // current ancestor's peer-suffix construction can use a
-            // `name@version` PeerId — see [`build_peer_id`] for the
-            // cycle handling.
-            let tree_node = &self.tree.dependencies_tree[node_id];
-            let pkg = &self.tree.packages[&tree_node.resolved_package_id];
-            return NodeOutput {
-                dep_path: DepPath::from(pkg.id.clone()),
-                external_resolved_peers: Arc::clone(&self.empty_resolved_peers),
-                auto_install_resolved_peers: HashMap::default(),
-                missing_peers: Arc::clone(&self.empty_missing_peers),
-                subtree_missing_by_pkg: None,
-            };
-        }
-        self.in_progress.insert(node_id.clone());
-
-        let fast_cached = {
-            let tree_node = &self.tree.dependencies_tree[node_id];
-            (tree_node.locked_peer_names.is_none() && tree_node.locked_peer_context.is_none())
-                .then(|| {
-                    self.find_fast_hit(node_id, parent_parent_refs, &tree_node.resolved_package_id)
-                })
-                .flatten()
-                .map(PeersCacheItem::to_cached_node_output)
-        };
-        if let Some(cached) = fast_cached {
-            let tree_node_depth = self.tree.dependencies_tree[node_id].depth;
-            return self.finish_cache_hit(
-                cached,
-                CacheHitContext {
-                    node_id,
-                    tree_node_depth,
-                    parent_chain_names,
-                    parent_pkg_ids_chain,
-                    preview_undo: None,
-                },
-            );
-        }
-        let (pkg_id, tree_node_depth, tree_node_installable, locked_peer_names) = {
-            let tree_node = &self.tree.dependencies_tree[node_id];
-            (
-                tree_node.resolved_package_id.clone(),
-                tree_node.depth,
-                tree_node.installable,
-                tree_node.locked_peer_names.clone(),
-            )
-        };
-        let pkg = self.owned_package(&pkg_id);
-        let (provider_children, preview_undo) = self.preview_peer_provider_children(node_id);
-        let mut refs_changed = locked_peer_names.is_some();
-        let parent_parent_refs = if let Some(locked_peer_names) = &locked_peer_names {
-            Arc::new(scoped_hoisted_optional_parent_refs(
-                parent_parent_refs,
-                locked_peer_names,
-                &self.opts.hoisted_optional_peer_node_ids,
-            ))
-        } else {
-            Arc::clone(parent_parent_refs)
-        };
-        let (pkg_name, _pkg_version) = pkg_name_version(&pkg.result);
-
-        // Build the ParentRefs map that descendants of this node see:
-        // parent's view + this node's own peer-relevant children. Kept
-        // behind `Arc` copy-on-write: most nodes contribute nothing, so
-        // they pass the parent's map down by refcount instead of cloning
-        // it — the per-node map clones dominated the walker's CPU time on
-        // peer-heavy workspaces.
-        let mut child_parent_refs = parent_parent_refs;
-        let mut new_parent_refs = ParentRefs::default();
-        for (alias, child_node_id) in &provider_children {
-            let Some(child_tree) = self.tree.dependencies_tree.get(child_node_id) else { continue };
-            let Some(child_pkg) = self.tree.packages.get(&child_tree.resolved_package_id) else {
-                continue;
-            };
-            insert_parent_ref(
-                &mut new_parent_refs,
-                alias,
-                child_node_id.clone(),
-                child_pkg,
-                child_tree.depth,
-            );
-        }
-        if !new_parent_refs.is_empty() {
-            refs_changed = true;
-            let refs = Arc::make_mut(&mut child_parent_refs);
-            // Built only when a name collision actually consults it — the
-            // common no-collision node never pays for the extra map clone.
-            let mut refs_with_new: Option<ParentRefs> = None;
-            for (name, mut new_parent_ref) in new_parent_refs.clone() {
-                if let Some(existing) = refs.get(&name) {
-                    let with_new = refs_with_new.get_or_insert_with(|| {
-                        let mut with_new = refs.clone();
-                        with_new.extend(new_parent_refs.clone());
-                        with_new
-                    });
-                    if !self.parent_refs_match(existing, &new_parent_ref)
-                        || self.inherited_parent_pkg_breaks_peer_diamond(
-                            with_new,
-                            existing,
-                            &new_parent_ref,
-                            node_id,
-                        )
-                    {
-                        new_parent_ref.occurrence = existing.occurrence + 1;
-                        refs.insert(name, new_parent_ref);
-                    }
-                } else {
-                    refs.insert(name, new_parent_ref);
-                }
-            }
-        }
-
-        let locked_pins =
-            self.locked_peer_context_pins(node_id, &pkg, &child_parent_refs, parent_node_ids);
-        if !locked_pins.is_empty() {
-            refs_changed = true;
-            let refs = Arc::make_mut(&mut child_parent_refs);
-            for (name, parent_ref) in locked_pins {
-                refs.insert(name, parent_ref);
-            }
-        }
-
-        // Record this node's parent context for the descendants'
-        // [`peers_cache`] lookups. We compute and store the snapshot
-        // before recursing so a cycle re-entry on a child also has
-        // access to its caller's parent context. Unchanged refs reuse the
-        // caller's snapshot instead of rebuilding an identical map.
-        let parent_dep_paths = if refs_changed {
-            self.parent_dep_paths_from_refs(&child_parent_refs)
-        } else {
-            Arc::clone(parent_dep_paths)
-        };
-        for child_node_id in
-            new_parent_refs.values().filter_map(|parent_ref| parent_ref.node_id.as_ref())
-        {
-            self.parent_pkgs_of_node.insert(child_node_id.clone(), Arc::clone(&parent_dep_paths));
-        }
-
-        // `peersCache` lookup. When an earlier walk of this same
-        // `pkgIdWithPatchHash` produced a result whose resolved-peer
-        // map and missing-peer set are compatible with the current
-        // parent peer context, reuse the cached `depPath` and external
-        // peer/missing maps without recursing.
-        //
-        // The cache lookup uses `child_parent_refs` (the augmented
-        // view) because a node's own children count as parents for
-        // its own descendants' peer resolution.
-        let cached =
-            self.find_hit(&child_parent_refs, &pkg.id).map(PeersCacheItem::to_cached_node_output);
-        if let Some(cached) = cached {
-            return self.finish_cache_hit(
-                cached,
-                CacheHitContext {
-                    node_id,
-                    tree_node_depth,
-                    parent_chain_names,
-                    parent_pkg_ids_chain,
-                    preview_undo,
-                },
-            );
-        }
-        let discovery_children = if self.discovery {
-            let node = &self.tree.dependencies_tree[node_id];
-            if let TreeChildren::Lazy { parent_ids } = &node.children {
-                Some((
-                    self.tree.children_by_id.get(&pkg.id).cloned().unwrap_or_default(),
-                    parent_ids.pushed(pkg.id.clone()),
-                ))
-            } else {
-                None
-            }
-        } else {
-            None
-        };
-        let (children_map, realize_undo) = if discovery_children.is_some() {
-            (BTreeMap::new(), None)
-        } else {
-            self.realize_children_with(node_id, Some(&provider_children))
-        };
-        let realize_undo = merge_realize_undo(preview_undo, realize_undo);
-        let current_parent_node_ids = parent_node_ids.pushed(node_id.clone());
-        let child_parent_pkg_ids_chain = if parent_pkg_ids_chain.contains(&pkg.id) {
-            parent_pkg_ids_chain.clone()
-        } else {
-            parent_pkg_ids_chain.pushed(pkg.id.clone())
-        };
-        let child_chain_names = parent_chain_names.pushed(pkg_name.clone());
-
-        // Recurse into children first (post-order). Discovery walks lazy
-        // children directly so cache hits never need occurrence-tree nodes.
-        let mut child_outputs = ChildOutputs::default();
-        if let Some((children, parent_ids)) = &discovery_children {
-            let child_aliases = ChildAliases::Deferred(children);
-            for repeated in [true, false] {
-                for edge in children.iter() {
-                    if child_parent_refs.contains_key(&edge.alias) != repeated
-                        || parent_ids.forms_cycle(&pkg.id, &edge.pkg_id)
-                    {
-                        continue;
-                    }
-                    let child_node_id =
-                        provider_children.get(&edge.alias).cloned().unwrap_or_else(|| {
-                            if self.tree.packages.get(&edge.pkg_id).is_some_and(|pkg| pkg.is_leaf) {
-                                NodeId::leaf(&edge.pkg_id)
-                            } else {
-                                NodeId::next()
-                            }
-                        });
-                    let child_output = if self.tree.dependencies_tree.contains_key(&child_node_id) {
-                        self.resolve_node(
-                            &child_node_id,
-                            &child_parent_refs,
-                            &parent_dep_paths,
-                            &child_chain_names,
-                            &current_parent_node_ids,
-                            &child_parent_pkg_ids_chain,
-                        )
-                    } else {
-                        self.resolve_deferred_child(DeferredChildContext {
-                            edge,
-                            node_id: child_node_id,
-                            parent_ids,
-                            parent_refs: &child_parent_refs,
-                            parent_dep_paths: &parent_dep_paths,
-                            chain_names: &child_chain_names,
-                            parent_node_ids: &current_parent_node_ids,
-                            parent_pkg_ids: &child_parent_pkg_ids_chain,
-                            depth: tree_node_depth + 1,
-                        })
-                    };
-                    child_outputs.push(&edge.alias, child_output, &child_aliases, false);
-                }
-            }
-        } else {
-            let child_aliases = ChildAliases::Realized(&children_map);
-            for repeated in [true, false] {
-                for (alias, child_node_id) in &children_map {
-                    if child_parent_refs.contains_key(alias) != repeated {
-                        continue;
-                    }
-                    let child_output = self.resolve_node(
-                        child_node_id,
-                        &child_parent_refs,
-                        &parent_dep_paths,
-                        &child_chain_names,
-                        &current_parent_node_ids,
-                        &child_parent_pkg_ids_chain,
-                    );
-                    child_outputs.push(alias, child_output, &child_aliases, !self.discovery);
-                }
-            }
-        }
-        let ChildOutputs {
-            external_peers: external_from_children,
-            mut auto_install_resolved_peers,
-            missing_peers: missing_from_children,
-            dep_paths: child_dep_paths,
-            mut missing_summaries,
-        } = child_outputs;
-
-        // Resolve this node's own peer requirements against the augmented
-        // ParentRefs visible at this node, including peer-relevant children.
-        let mut own_resolved_peers: HashMap<String, NodeId> = HashMap::default();
-        let mut own_missing_peers: HashMap<String, MissingPeerInfo> = HashMap::default();
-        for (peer_name, peer_dep) in &pkg.peer_dependencies {
-            self.resolve_one_peer(
-                &pkg_name,
-                peer_name,
-                peer_dep,
-                &child_parent_refs,
-                &child_chain_names,
-                parent_pkg_ids_chain,
-                &mut own_resolved_peers,
-                &mut own_missing_peers,
-            );
-            if let Some(peer_node_id) = own_resolved_peers.get(peer_name) {
-                auto_install_resolved_peers.insert(peer_name.clone(), peer_node_id.clone());
-            }
-        }
-
-        // Combine all resolved peers (this node's own + descendants').
-        // Filter out the node's own name (a package doesn't peer-depend
-        // on itself).
-        let mut all_resolved_peers = external_from_children;
-        for (peer_alias, peer_node_id) in &own_resolved_peers {
-            all_resolved_peers.insert(peer_alias.clone(), peer_node_id.clone());
-        }
-        all_resolved_peers.remove(&pkg_name);
-
-        // Same for missing peers (children + own).
-        let mut all_missing_peers = missing_from_children.clone();
-        for (peer_alias, info) in &own_missing_peers {
-            all_missing_peers.insert(peer_alias.clone(), info.clone());
-        }
-
-        // Construct the depPath. Empty resolved-peers ⇒ pure node:
-        // depPath = pkgIdWithPatchHash.
-        let dep_path = if all_resolved_peers.is_empty() {
-            DepPath::from(pkg.id.clone())
-        } else {
-            let peer_ids: Vec<PeerId> = all_resolved_peers
-                .iter()
-                .map(|(peer_alias, peer_node_id)| self.build_peer_id(peer_alias, peer_node_id))
-                .collect();
-            let suffix = create_peer_dep_graph_hash(&peer_ids, self.opts.peers_suffix_max_length);
-            DepPath::from(format!("{}{}", pkg.id, suffix))
-        };
-
-        // Register the depPath ↔ NodeId mapping and per-node
-        // propagated state before inserting into the graph (so any
-        // cycle the graph insert hits via `child_dep_paths` can find
-        // this node's depPath).
-        self.remember_resolved_node(node_id, &dep_path);
-
-        let own_missing = (!missing_from_children.is_empty())
-            .then(|| (pkg.id.clone(), missing_from_children.keys().cloned().collect()));
-        let subtree_missing_by_pkg = match (own_missing, missing_summaries.len()) {
-            (None, 0) => None,
-            (None, 1) => missing_summaries.pop(),
-            (own, _) => Some(Arc::new(MissingSummary { own, children: missing_summaries })),
-        };
-
-        let is_pure = all_resolved_peers.is_empty() && all_missing_peers.is_empty();
-        let all_resolved_peers = Arc::new(all_resolved_peers);
-        let all_missing_peers = Arc::new(all_missing_peers);
-        let missing_from_children = Arc::new(missing_from_children);
-
-        if !self.discovery {
-            self.node_external_peers.insert(node_id.clone(), Arc::clone(&all_resolved_peers));
-            self.node_missing_peers.insert(node_id.clone(), Arc::clone(&all_missing_peers));
-            self.node_missing_peers_of_children
-                .insert(node_id.clone(), Arc::clone(&missing_from_children));
-        }
-
-        // Record this walk's outcome in the per-`pkgIdWithPatchHash`
-        // caches. Pure subtrees go in [`Self::pure_pkgs`] for the
-        // fast-path early return at the top of [`resolve_node`];
-        // non-pure subtrees push a [`PeersCacheItem`] so a future
-        // visit with a compatible parent context can short-circuit
-        // via [`Self::find_hit`].
-        //
-        // A cycle re-entry resolves against truncated children (the cycle is
-        // broken by dropping the repeated package's subtree), so its empty or
-        // partial peer sets are not authoritative for the package as a whole.
-        // Caching that verdict would let it short-circuit other occurrences
-        // that can see the full subtree, dropping their
-        // `transitivePeerDependencies` depending on traversal order and
-        // churning the lockfile (https://github.com/pnpm/pnpm/issues/5108).
-        let resolved_through_cycle = parent_pkg_ids_chain.contains(&pkg.id);
-        if resolved_through_cycle {
-            // Leave both caches untouched so later occurrences re-resolve (or
-            // hit the authoritative entry of the same package) instead of
-            // reusing this partial one.
-        } else if is_pure {
-            self.pure_pkgs.insert(pkg.id.clone(), dep_path.clone());
-        } else {
-            self.retained_peer_node_ids.extend(all_resolved_peers.values().cloned());
-            self.peers_cache.entry(pkg.id.clone()).or_default().push(PeersCacheItem {
-                dep_path: dep_path.clone(),
-                resolved_peers: Arc::clone(&all_resolved_peers),
-                missing_peers: Arc::clone(&all_missing_peers),
-                missing_peers_of_children: Arc::clone(&missing_from_children),
-                subtree_missing_by_pkg: subtree_missing_by_pkg.clone(),
-            });
-        }
-
-        if !self.discovery {
-            // The children's depPath edges become this node's graph children.
-            // Resolved peers become extra edges, aliased by peer name. If a
-            // peer's depPath isn't known yet — typically a later sibling
-            // direct dep — defer the edge to the post-walk patch pass; the
-            // install layer drives off `graph_children`, so skipping the
-            // edge entirely would leave the peer un-symlinked in the
-            // parent's slot.
-            let mut graph_children = BTreeMap::new();
-            for (alias, child_node_id) in
-                self.previously_resolved_children(parent_node_ids, parent_pkg_ids_chain, &pkg.id)
-            {
-                self.add_graph_child_or_pending(
-                    &mut graph_children,
-                    &dep_path,
-                    alias,
-                    child_node_id,
-                );
-            }
-            for (alias, child_dep_path) in child_dep_paths {
-                graph_children.insert(alias, child_dep_path);
-            }
-            for (peer_alias, peer_node_id) in all_resolved_peers.iter() {
-                self.add_graph_child_or_pending(
-                    &mut graph_children,
-                    &dep_path,
-                    peer_alias.clone(),
-                    peer_node_id.clone(),
-                );
-            }
-
-            // Compute transitive peer set: peers visible in this subtree
-            // that are NOT declared in this package's own peerDependencies.
-            let mut transitive_peer_dependencies: HashSet<String> = HashSet::default();
-            for peer_alias in all_resolved_peers.keys() {
-                if !pkg.peer_dependencies.contains_key(peer_alias) {
-                    transitive_peer_dependencies.insert(peer_alias.clone());
-                }
-            }
-            for peer_alias in all_missing_peers.keys() {
-                if !pkg.peer_dependencies.contains_key(peer_alias) {
-                    transitive_peer_dependencies.insert(peer_alias.clone());
-                }
-            }
-
-            // Capture this node's NodeId-level edges + metadata for the
-            // post-walk [`Walker::build_final_dep_paths`] rebuild. Edges are
-            // the node's regular children overlaid with its *own* resolved
-            // peers — this node's own peer resolution, not the descendants'
-            // peers bubbled up for the suffix. A peer a descendant resolved
-            // (e.g. `debug`'s optional `supports-color`) is symlinked at the
-            // descendant that declares it, so it must not appear in this
-            // node's dependencies. Carries NodeIds so the rebuild can
-            // resolve each to its corrected final depPath.
-            let mut record_edges =
-                self.previously_resolved_children(parent_node_ids, parent_pkg_ids_chain, &pkg.id);
-            record_edges.extend(children_map.clone());
-            for (peer_alias, peer_node_id) in &own_resolved_peers {
-                record_edges.insert(peer_alias.clone(), peer_node_id.clone());
-            }
-            let optional_child_aliases = self.optional_child_aliases(&pkg.id, &record_edges);
-            let record_order = self.next_record_order;
-            self.next_record_order += 1;
-            self.node_records.insert(
-                node_id.clone(),
-                NodeRecord {
-                    edges: record_edges,
-                    optional_child_aliases: optional_child_aliases.clone(),
-                    transitive_peer_dependencies: transitive_peer_dependencies.clone(),
-                    depth: tree_node_depth,
-                    installable: tree_node_installable,
-                    is_pure,
-                    order: record_order,
-                },
-            );
-
-            // Multiple visits with the same depPath collapse onto the same
-            // graph entry. On a conflict, keep the entry with the smallest
-            // `depth` so install order matches.
-            self.graph
-                .entry(dep_path.clone())
-                .and_modify(|node| {
-                    if node.depth > tree_node_depth {
-                        node.depth = tree_node_depth;
-                    }
-                })
-                .or_insert(DependenciesGraphNode {
-                    dep_path: dep_path.clone(),
-                    resolved_package_id: pkg.id.clone(),
-                    resolve_result: Arc::clone(&pkg.result),
-                    children: graph_children,
-                    optional_children: optional_child_aliases,
-                    peer_dependencies: pkg.peer_dependencies.clone(),
-                    transitive_peer_dependencies,
-                    resolved_peer_names: all_resolved_peers.keys().cloned().collect(),
-                    depth: tree_node_depth,
-                    installable: tree_node_installable,
-                    is_pure,
-                    optional: pkg.optional,
-                });
-        }
-
-        self.in_progress.remove(node_id);
-
-        let external_to_report: HashMap<String, NodeId> = all_resolved_peers
-            .iter()
-            .filter(|(peer_alias, _)| {
-                !children_map.contains_key(peer_alias.as_str())
-                    && discovery_children.as_ref().is_none_or(|(children, _)| {
-                        !children.iter().any(|edge| edge.alias == **peer_alias)
-                    })
-            })
-            .map(|(peer_alias, peer_node_id)| (peer_alias.clone(), peer_node_id.clone()))
-            .collect();
-
-        let output = NodeOutput {
-            dep_path,
-            external_resolved_peers: Arc::new(external_to_report),
-            auto_install_resolved_peers,
-            missing_peers: all_missing_peers,
-            subtree_missing_by_pkg,
-        };
-        if self.discovery {
-            self.undo_realize(node_id, realize_undo, Some(&output));
-        }
-        output
-    }
-
-    fn finish_cache_hit(
-        &mut self,
-        cached: CachedNodeOutput,
-        context: CacheHitContext<'_>,
-    ) -> NodeOutput {
-        let CacheHitContext {
-            node_id,
-            tree_node_depth,
-            parent_chain_names,
-            parent_pkg_ids_chain,
-            preview_undo,
-        } = context;
-        let output = cached.output;
-        self.undo_realize(node_id, preview_undo, None);
-
-        if !output.missing_peers.is_empty() {
-            let pkg_id = self.tree.dependencies_tree[node_id].resolved_package_id.clone();
-            let chain_with_self = parent_pkg_ids_chain.pushed(pkg_id.clone());
-            let pkg_name = pkg_name_version(&self.tree.packages[&pkg_id].result).0;
-            for (peer_name, info) in output.missing_peers.iter() {
-                if self.missing_issue_suppressed(&chain_with_self, peer_name) {
-                    continue;
-                }
-                self.record_missing_issue(
-                    peer_name,
-                    MissingPeer {
-                        wanted_range: get_peer_version_range(&info.range),
-                        raw_range: info.range.clone(),
-                        optional: info.optional,
-                        parents: self.issue_parents(parent_chain_names, &pkg_name),
-                    },
-                    &chain_with_self,
-                );
-            }
-        }
-        self.remember_resolved_node(node_id, &output.dep_path);
-        if !self.discovery {
-            self.node_external_peers
-                .insert(node_id.clone(), Arc::clone(&output.external_resolved_peers));
-            self.node_missing_peers.insert(node_id.clone(), Arc::clone(&output.missing_peers));
-            self.node_missing_peers_of_children
-                .insert(node_id.clone(), cached.missing_peers_of_children);
-        }
-        if let Some(node) = self.graph.get_mut(&output.dep_path)
-            && node.depth > tree_node_depth
-        {
-            node.depth = tree_node_depth;
-        }
-        self.in_progress.remove(node_id);
-        output
-    }
-
-    /// `true` when a missing-peer issue for `peer_name` under the
-    /// given ancestor chain must not be emitted for the hoist input.
-    /// See [`ResolvePeersOptions::hoist_missing_scope`].
-    /// The upstream locked-peer-provider reuse block
-    /// (`resolvePeers.ts:594`): for each `peer name → provider DepPath`
-    /// the wanted lockfile recorded on this node, re-pin the provider
-    /// into `parent_refs` when it is still reachable in the current
-    /// tree, resolved to the same path in the previous pass, carries no
-    /// peer suffix of its own, has not diverged in this pass, is not
-    /// overridden by a current provider that must win, and satisfies
-    /// the node's current peer range.
-    /// The pins [`apply_locked_peer_context` upstream] would insert into
-    /// `parent_refs`, computed without mutating it so the caller can keep
-    /// sharing an unchanged map. Each pin reads only its own name's
-    /// current binding, so collecting against the pre-pin map is
-    /// equivalent to inserting while iterating.
-    fn locked_peer_context_pins(
-        &self,
-        node_id: &NodeId,
-        pkg: &ResolvedPackage,
-        parent_refs: &ParentRefs,
-        parent_node_ids: &SharedChain<NodeId>,
-    ) -> Vec<(String, ParentRef)> {
-        let mut pins = Vec::new();
-        let (Some(locked_peer_context), Some(provider_paths)) = (
-            self.tree
-                .dependencies_tree
-                .get(node_id)
-                .and_then(|tree_node| tree_node.locked_peer_context.as_ref()),
-            self.opts.resolved_peer_provider_paths.as_ref(),
-        ) else {
-            return pins;
-        };
-        for (peer_name, previous_dep_path) in locked_peer_context {
-            let Some(peer_node_id) = self.node_ids_by_previous_dep_path.get(previous_dep_path)
-            else {
-                continue;
-            };
-            let Some(peer_dep) = pkg.peer_dependencies.get(peer_name) else { continue };
-            if provider_paths.get(peer_node_id) != Some(previous_dep_path) {
-                continue;
-            }
-            // Only pin providers that have no peer context of their
-            // own — a suffixed path depends on the very bindings this
-            // pass is still computing.
-            if index_of_dep_path_suffix(previous_dep_path.as_str()).peers_index.is_some() {
-                continue;
-            }
-            // A provider that already resolved to a different path
-            // this pass must not be rebound.
-            if self
-                .node_dep_paths
-                .get(peer_node_id)
-                .is_some_and(|current| current != previous_dep_path)
-            {
-                continue;
-            }
-            if self.has_current_peer_provider_that_must_win(peer_name, parent_refs, parent_node_ids)
-            {
-                continue;
-            }
-            let Some(peer_tree_node) = self.tree.dependencies_tree.get(peer_node_id) else {
-                continue;
-            };
-            let Some(peer_pkg) = self.tree.packages.get(&peer_tree_node.resolved_package_id) else {
-                continue;
-            };
-            let (_, peer_version) = pkg_name_version(&peer_pkg.result);
-            if !satisfies_with_prereleases(
-                &peer_version,
-                &get_peer_version_range(&peer_dep.version),
-            ) {
-                continue;
-            }
-            // Upstream builds the pinned ref through `toPkgByName`,
-            // which always starts at occurrence 0; the shadow counter
-            // only tracks child-level replacements.
-            pins.push((
-                peer_name.clone(),
-                ParentRef {
-                    version: peer_version,
-                    node_id: Some(peer_node_id.clone()),
-                    alias: Some(peer_name.clone()),
-                    depth: peer_tree_node.depth,
-                    occurrence: 0,
-                },
-            ));
-        }
-        pins
-    }
-
-    /// The upstream `hasCurrentPeerProviderThatMustWin`: the current
-    /// provider bound for `peer_name` wins over a locked one when it is
-    /// an importer direct dep under a *different* alias, one the user
-    /// explicitly requested, one the manifest declares with no
-    /// wanted-lockfile resolution, or a child an ancestor re-resolved
-    /// away from the lockfile.
-    fn has_current_peer_provider_that_must_win(
-        &self,
-        peer_name: &str,
-        parent_refs: &ParentRefs,
-        parent_node_ids: &SharedChain<NodeId>,
-    ) -> bool {
-        let Some(peer_node_id) =
-            parent_refs.get(peer_name).and_then(|parent| parent.node_id.as_ref())
-        else {
-            return false;
-        };
-        for source in &self.current_provider_sources {
-            for (alias, direct_node_id) in &source.direct_node_ids_by_alias {
-                if direct_node_id == peer_node_id
-                    && (alias != peer_name
-                        || source.explicitly_requested_direct_dependencies.contains(alias)
-                        || (source.declared_direct_dependencies.contains(alias)
-                            && self
-                                .tree
-                                .dependencies_tree
-                                .get(peer_node_id)
-                                .is_none_or(|node| node.previous_dep_path.is_none())))
-                {
-                    return true;
-                }
-            }
-        }
-        for parent_node_id in parent_node_ids.iter() {
-            let Some(parent_node) = self.tree.dependencies_tree.get(parent_node_id) else {
-                continue;
-            };
-            let Some(must_win) =
-                parent_node.dependency_names_whose_current_provider_must_win.as_ref()
-            else {
-                continue;
-            };
-            // Ancestors on the walk path always have realized children.
-            let TreeChildren::Realized(children) = &parent_node.children else { continue };
-            if must_win.iter().any(|alias| children.get(alias) == Some(peer_node_id)) {
-                return true;
-            }
-        }
-        false
-    }
-
-    fn missing_issue_suppressed(
-        &self,
-        ancestor_pkg_ids: &SharedChain<String>,
-        peer_name: &str,
-    ) -> bool {
-        let Some(scope) = self.opts.hoist_missing_scope.as_ref() else { return false };
-        scope.suppresses_iter(ancestor_pkg_ids.iter(), peer_name)
-    }
-
-    fn record_missing_issue(
-        &mut self,
-        peer_name: &str,
-        issue: MissingPeer,
-        ancestor_pkg_ids: &SharedChain<String>,
-    ) {
-        self.issues.missing.entry(peer_name.to_string()).or_default().push(issue);
-        if self.discovery {
-            self.missing_ancestor_pkg_ids
-                .entry(peer_name.to_string())
-                .or_default()
-                .push(ancestor_pkg_ids.clone());
-        }
-    }
-
-    fn issue_parents(&self, chain: &SharedChain<String>, pkg_name: &str) -> Vec<ParentPackageRef> {
-        if self.discovery { Vec::new() } else { parents_from_chain(chain, pkg_name) }
-    }
-
-    #[expect(
-        clippy::too_many_arguments,
-        reason = "internal walker helper threading per-node context, mirrors the resolve_node parameter set"
-    )]
-    fn resolve_one_peer(
-        &mut self,
-        pkg_name: &str,
-        peer_name: &str,
-        peer_dep: &PeerDep,
-        parent_refs: &ParentRefs,
-        chain: &SharedChain<String>,
-        ancestor_pkg_ids: &SharedChain<String>,
-        resolved: &mut HashMap<String, NodeId>,
-        missing: &mut HashMap<String, MissingPeerInfo>,
-    ) {
-        let raw_range = peer_dep.version.as_str();
-        // The stored range keeps the original scheme (only `workspace:` is
-        // stripped) so it still selects the package to auto-install for a
-        // missing peer, e.g. `work:5.x.x` fetches from the `work` registry.
-        let range_for_match = raw_range.strip_prefix("workspace:").unwrap_or(raw_range);
-        // The satisfaction check needs a comparable semver range, so
-        // named-registry/`npm:` bodies are extracted and opaque specs become `*`.
-        let range_for_satisfies = get_peer_version_range(raw_range);
-        let optional = peer_dep.optional;
-
-        match parent_refs.get(peer_name) {
-            None => {
-                missing.insert(
-                    peer_name.to_string(),
-                    MissingPeerInfo { range: range_for_match.to_string(), optional },
-                );
-                if !self.missing_issue_suppressed(ancestor_pkg_ids, peer_name) {
-                    self.record_missing_issue(
-                        peer_name,
-                        MissingPeer {
-                            wanted_range: range_for_satisfies,
-                            raw_range: range_for_match.to_string(),
-                            optional,
-                            parents: self.issue_parents(chain, pkg_name),
-                        },
-                        ancestor_pkg_ids,
-                    );
-                }
-            }
-            Some(parent) => {
-                if !satisfies_with_prereleases(&parent.version, &range_for_satisfies) {
-                    let parents = self.issue_parents(chain, pkg_name);
-                    self.issues.bad.entry(peer_name.to_string()).or_default().push(
-                        PeerDependencyIssue {
-                            wanted_range: range_for_satisfies,
-                            found_version: parent.version.clone(),
-                            optional,
-                            parents,
-                            resolved_from: Vec::new(),
-                        },
-                    );
-                }
-                if let Some(parent_node_id) = parent.node_id.as_ref() {
-                    resolved.insert(peer_name.to_string(), parent_node_id.clone());
-                }
-            }
-        }
-    }
-
-    /// Build the [`PeerId`] contribution for one resolved peer.
-    ///
-    /// Precedence:
-    ///
-    /// 1. **`link:<rel>` `NodeIds`** — emit
-    ///    `PeerId::Pair { name: peer_alias, version: link_path_to_peer_version(rel) }`
-    ///    so the peer-suffix segment reads as `name@encoded_path`
-    ///    instead of carrying the raw link target. This branch fires
-    ///    for both workspace-link parents and the
-    ///    `excludeLinksFromLockfile` remap that points the parent at
-    ///    `link:node_modules/<alias>`.
-    /// 2. **`dedupe_peers` enabled** — emit `name@version` from the
-    ///    resolved package so recursive peer suffixes collapse like
-    ///    `(foo@1.0.0(bar@2.0.0))` → `(foo@1.0.0)`.
-    /// 3. **The peer's `DepPath`** once it has been walked —
-    ///    `node_dep_paths` lookup, emitted as [`PeerId::DepPath`].
-    /// 4. **Cycle fallback** — `name@version` from the resolved package,
-    ///    emitted as [`PeerId::Pair`].
-    fn build_peer_id(&self, peer_alias: &str, peer_node_id: &NodeId) -> PeerId {
-        if let NodeId::Leaf(id) = peer_node_id
-            && let Some(rel) = id.strip_prefix("link:")
-        {
-            return PeerId::Pair {
-                name: peer_alias.to_string(),
-                version: link_path_to_peer_version(rel),
-            };
-        }
-        if self.opts.dedupe_peers
-            && let Some(tree_node) = self.tree.dependencies_tree.get(peer_node_id)
-            && let Some(pkg) = self.tree.packages.get(&tree_node.resolved_package_id)
-        {
-            let (name, version) = pkg_name_version(&pkg.result);
-            return PeerId::Pair { name, version };
-        }
-        if let Some(dep_path) = self.node_dep_paths.get(peer_node_id) {
-            return PeerId::DepPath(dep_path.clone());
-        }
-        let tree_node = &self.tree.dependencies_tree[peer_node_id];
-        let pkg = &self.tree.packages[&tree_node.resolved_package_id];
-        let (name, version) = pkg_name_version(&pkg.result);
-        PeerId::Pair { name, version }
-    }
-
-    /// Resolve `node_id` to the depPath computed during the main walk.
-    /// Peerless nodes are already final at this point; nodes with peers
-    /// may still be missing a pending peer provider's own final suffix.
-    fn provisional_dep_path_of(&self, node_id: &NodeId) -> DepPath {
-        if let Some(dep_path) = self.node_dep_paths.get(node_id) {
-            return dep_path.clone();
-        }
-        if let Some(dep_path) = link_node_id_as_dep_path(node_id) {
-            return dep_path;
-        }
-        let pkg_id = &self.tree.dependencies_tree[node_id].resolved_package_id;
-        DepPath::from(self.tree.packages[pkg_id].id.clone())
-    }
-
-    /// Resolve `node_id` to the depPath the rebuilt graph should key /
-    /// reference it by. Prefers the corrected `final_dep_paths` entry
-    /// and falls back to the provisional value for peerless nodes.
-    fn final_dep_path_of(
-        &self,
-        node_id: &NodeId,
-        final_dep_paths: &HashMap<NodeId, DepPath>,
-    ) -> DepPath {
-        if let Some(dep_path) = final_dep_paths.get(node_id) {
-            return dep_path.clone();
-        }
-        self.provisional_dep_path_of(node_id)
-    }
-
-    /// One resolved-peer slot of `node_id`'s suffix, computed against the
-    /// already-finalized depPaths. Like [`Self::build_peer_id`] but
-    /// substitutes the provisional cycle fallback with a strongly-
-    /// connected-component test: a peer is collapsed to `name@version`
-    /// only when it shares a peer-graph SCC with `node_id` (a genuine
-    /// cycle). Non-cyclic peers carry their full depPath.
-    fn final_peer_id(
-        &self,
-        node_id: &NodeId,
-        peer_alias: &str,
-        peer_node_id: &NodeId,
-        context: FinalPeerContext<'_>,
-        final_dep_paths: &mut HashMap<NodeId, DepPath>,
-        visiting: &mut HashSet<NodeId>,
-    ) -> PeerId {
-        if let NodeId::Leaf(id) = peer_node_id
-            && let Some(rel) = id.strip_prefix("link:")
-        {
-            return PeerId::Pair {
-                name: peer_alias.to_string(),
-                version: link_path_to_peer_version(rel),
-            };
-        }
-        let pair = || {
-            let tree_node = &self.tree.dependencies_tree[peer_node_id];
-            let pkg = &self.tree.packages[&tree_node.resolved_package_id];
-            let (name, version) = pkg_name_version(&pkg.result);
-            PeerId::Pair { name, version }
-        };
-        if self.opts.dedupe_peers && self.tree.dependencies_tree.contains_key(peer_node_id) {
-            return pair();
-        }
-        if context
-            .scc_of
-            .get(node_id)
-            .is_some_and(|node_scc| context.scc_of.get(peer_node_id) == Some(node_scc))
-        {
-            return pair();
-        }
-        if context.cyclic_peer_names.contains(peer_alias) {
-            return pair();
-        }
-        PeerId::DepPath(self.final_dep_path_for_node(
-            peer_node_id,
-            context.scc_of,
-            context.cyclic_peer_names,
-            final_dep_paths,
-            visiting,
-        ))
-    }
-
-    /// Recompute every node's depPath with its resolved peers' *full*
-    /// suffixes. Genuine peer cycles (detected as multi-node peer-graph
-    /// SCCs, or self-loops) keep the `name@version` collapse; every other
-    /// peer slot carries the peer's own depPath. The cycle detection
-    /// runs synchronously over the already-walked graph.
-    fn build_final_dep_paths(&self) -> HashMap<NodeId, DepPath> {
-        let (_, scc_of) = self.peer_sccs();
-        let cyclic_peer_names = self.cyclic_peer_names();
-        let mut final_dep_paths: HashMap<NodeId, DepPath> = HashMap::default();
-        let mut visiting = HashSet::default();
-        let mut node_ids: Vec<NodeId> = self.node_external_peers.keys().cloned().collect();
-        node_ids.sort_by_key(node_id_sort_key);
-        for node_id in node_ids {
-            self.final_dep_path_for_node(
-                &node_id,
-                &scc_of,
-                &cyclic_peer_names,
-                &mut final_dep_paths,
-                &mut visiting,
-            );
-        }
-        final_dep_paths
-    }
-
-    fn final_dep_path_for_node(
-        &self,
-        node_id: &NodeId,
-        scc_of: &HashMap<NodeId, usize>,
-        cyclic_peer_names: &HashSet<String>,
-        final_dep_paths: &mut HashMap<NodeId, DepPath>,
-        visiting: &mut HashSet<NodeId>,
-    ) -> DepPath {
-        if let Some(dep_path) = final_dep_paths.get(node_id) {
-            return dep_path.clone();
-        }
-        let Some(peers) = self.node_external_peers.get(node_id) else {
-            return self.provisional_dep_path_of(node_id);
-        };
-        if peers.is_empty() {
-            return self.provisional_dep_path_of(node_id);
-        }
-        if !visiting.insert(node_id.clone()) {
-            return self.provisional_dep_path_of(node_id);
-        }
-        let peer_ids: Vec<PeerId> = peers
-            .iter()
-            .map(|(peer_alias, peer_node_id)| {
-                self.final_peer_id(
-                    node_id,
-                    peer_alias,
-                    peer_node_id,
-                    FinalPeerContext { scc_of, cyclic_peer_names },
-                    final_dep_paths,
-                    visiting,
-                )
-            })
-            .collect();
-        let suffix = create_peer_dep_graph_hash(&peer_ids, self.opts.peers_suffix_max_length);
-        let pkg_id = &self.tree.dependencies_tree[node_id].resolved_package_id;
-        let dep_path = DepPath::from(format!("{}{}", self.tree.packages[pkg_id].id, suffix));
-        final_dep_paths.insert(node_id.clone(), dep_path.clone());
-        visiting.remove(node_id);
-        dep_path
-    }
-
-    fn cyclic_peer_names(&self) -> HashSet<String> {
-        let mut graph: BTreeMap<String, BTreeSet<String>> = BTreeMap::new();
-        for (node_id, peers) in &self.node_external_peers {
-            if peers.is_empty() {
-                continue;
-            }
-            let tree_node = &self.tree.dependencies_tree[node_id];
-            let pkg = &self.tree.packages[&tree_node.resolved_package_id];
-            let (name, _) = pkg_name_version(&pkg.result);
-            let edges = graph.entry(name).or_default();
-            for peer_alias in peers.keys() {
-                edges.insert(peer_alias.clone());
-            }
-        }
-        let peer_names: Vec<String> =
-            graph.values().flat_map(|edges| edges.iter().cloned()).collect();
-        for peer_name in peer_names {
-            graph.entry(peer_name).or_default();
-        }
-
-        struct PeerNameTarjan<'a> {
-            graph: &'a BTreeMap<String, BTreeSet<String>>,
-            index_of: HashMap<String, u32>,
-            low_of: HashMap<String, u32>,
-            on_stack: HashSet<String>,
-            tarjan_stack: Vec<String>,
-            cyclic: HashSet<String>,
-            next_index: u32,
-        }
-
-        impl PeerNameTarjan<'_> {
-            fn strongconnect(&mut self, name: &str) {
-                self.index_of.insert(name.to_string(), self.next_index);
-                self.low_of.insert(name.to_string(), self.next_index);
-                self.next_index += 1;
-                self.on_stack.insert(name.to_string());
-                self.tarjan_stack.push(name.to_string());
-
-                if let Some(neighbors) = self.graph.get(name) {
-                    for child in neighbors {
-                        if !self.index_of.contains_key(child) {
-                            self.strongconnect(child);
-                            let name_low = self.low_of[name];
-                            let child_low = self.low_of[child];
-                            self.low_of.insert(name.to_string(), name_low.min(child_low));
-                        } else if self.on_stack.contains(child) {
-                            let name_low = self.low_of[name];
-                            let child_index = self.index_of[child];
-                            self.low_of.insert(name.to_string(), name_low.min(child_index));
-                        }
-                    }
-                }
-
-                if self.low_of[name] == self.index_of[name] {
-                    let mut component = Vec::new();
-                    while let Some(member) = self.tarjan_stack.pop() {
-                        self.on_stack.remove(&member);
-                        let is_root = member == name;
-                        component.push(member);
-                        if is_root {
-                            break;
-                        }
-                    }
-                    let self_loop = component.first().is_some_and(|member| {
-                        self.graph.get(member).is_some_and(|edges| edges.contains(member))
-                    });
-                    if component.len() > 1 || self_loop {
-                        self.cyclic.extend(component);
-                    }
-                }
-            }
-        }
-
-        let mut tarjan = PeerNameTarjan {
-            graph: &graph,
-            index_of: HashMap::default(),
-            low_of: HashMap::default(),
-            on_stack: HashSet::default(),
-            tarjan_stack: Vec::new(),
-            cyclic: HashSet::default(),
-            next_index: 0,
-        };
-        for name in graph.keys() {
-            if !tarjan.index_of.contains_key(name) {
-                tarjan.strongconnect(name);
-            }
-        }
-        tarjan.cyclic
-    }
-
-    /// Strongly-connected components of the peer graph (node → resolved
-    /// peers, restricted to peers that themselves carry peers — peerless
-    /// peers can't close a cycle). Iterative Tarjan, returning the SCCs
-    /// in reverse-topological order plus a `NodeId → SCC index` map.
-    fn peer_sccs(&self) -> (Vec<Vec<NodeId>>, HashMap<NodeId, usize>) {
-        let mut participants: Vec<NodeId> = self
-            .node_external_peers
-            .iter()
-            .filter(|(_, peers)| !peers.is_empty())
-            .map(|(node_id, _)| node_id.clone())
-            .collect();
-        participants.sort_by_key(node_id_sort_key);
-        participants.dedup();
-        let participant_set: HashSet<NodeId> = participants.iter().cloned().collect();
-        let neighbors = |node_id: &NodeId| -> Vec<NodeId> {
-            let mut out: Vec<NodeId> = self
-                .node_external_peers
-                .get(node_id)
-                .into_iter()
-                .flat_map(|peers| peers.values())
-                .filter(|peer| participant_set.contains(*peer))
-                .cloned()
-                .collect();
-            out.sort_by_key(node_id_sort_key);
-            out
-        };
-
-        let mut index_of: HashMap<NodeId, u32> = HashMap::default();
-        let mut low_of: HashMap<NodeId, u32> = HashMap::default();
-        let mut on_stack: HashSet<NodeId> = HashSet::default();
-        let mut tarjan_stack: Vec<NodeId> = Vec::new();
-        let mut sccs: Vec<Vec<NodeId>> = Vec::new();
-        let mut scc_of: HashMap<NodeId, usize> = HashMap::default();
-        let mut next_index: u32 = 0;
-
-        // Explicit DFS stack of (node, neighbors, cursor) so deep peer
-        // graphs don't overflow the call stack.
-        for root in &participants {
-            if index_of.contains_key(root) {
-                continue;
-            }
-            let mut work: Vec<(NodeId, Vec<NodeId>, usize)> =
-                vec![(root.clone(), neighbors(root), 0)];
-            while let Some((node_id, succ, cursor)) = work.last_mut() {
-                if *cursor == 0 {
-                    index_of.insert(node_id.clone(), next_index);
-                    low_of.insert(node_id.clone(), next_index);
-                    next_index += 1;
-                    on_stack.insert(node_id.clone());
-                    tarjan_stack.push(node_id.clone());
-                }
-                if *cursor < succ.len() {
-                    let child = succ[*cursor].clone();
-                    *cursor += 1;
-                    if !index_of.contains_key(&child) {
-                        let child_succ = neighbors(&child);
-                        work.push((child, child_succ, 0));
-                    } else if on_stack.contains(&child) {
-                        let node_low = low_of[node_id];
-                        let child_index = index_of[&child];
-                        low_of.insert(node_id.clone(), node_low.min(child_index));
-                    }
-                    continue;
-                }
-                // All successors visited — close this node.
-                let node_id = node_id.clone();
-                if low_of[&node_id] == index_of[&node_id] {
-                    let scc_index = sccs.len();
-                    let mut component = Vec::new();
-                    while let Some(member) = tarjan_stack.pop() {
-                        on_stack.remove(&member);
-                        scc_of.insert(member.clone(), scc_index);
-                        let is_root = member == node_id;
-                        component.push(member);
-                        if is_root {
-                            break;
-                        }
-                    }
-                    sccs.push(component);
-                }
-                work.pop();
-                if let Some((parent, _, _)) = work.last() {
-                    let parent_low = low_of[parent];
-                    let node_low = low_of[&node_id];
-                    low_of.insert(parent.clone(), parent_low.min(node_low));
-                }
-            }
-        }
-        (sccs, scc_of)
-    }
-
-    /// Rebuild the depPath-keyed graph from the per-`NodeId`
-    /// [`NodeRecord`]s using the corrected `final_dep_paths`. Nodes that
-    /// resolve to the same final depPath merge (taking the smallest
-    /// `depth`, like the inline build); nodes whose suffix was
-    /// previously collapsed by the cycle fallback now split into
-    /// distinct entries.
-    ///
-    /// Every edge — a regular child or a resolved peer — points at the
-    /// depPath the edge's own node resolved to, matching upstream's
-    /// `resolveChildren`, which maps each `childrenNodeIds` entry
-    /// through `pathsByNodeId`. A peer provider therefore keeps its own
-    /// peer suffix even where the consumer resolved none of those peers
-    /// itself.
-    fn build_final_graph(&self, final_dep_paths: &HashMap<NodeId, DepPath>) -> DependenciesGraph {
-        // Minimum tree depth across *every* occurrence that resolves to a
-        // given final depPath. `pure_pkgs` / `find_hit` revisits
-        // short-circuit before a [`NodeRecord`] is created, so iterating
-        // `node_records` alone would restore the first (possibly deeper)
-        // walk's depth and miss a later shallower revisit. `node_dep_paths`
-        // carries every walked NodeId, so recompute the `Math.min` depth
-        // tie-break here — the inline build threaded it through `self.graph`,
-        // which this rebuild discards.
-        let mut min_depth: HashMap<DepPath, i32> = HashMap::default();
-        for node_id in self.node_dep_paths.keys() {
-            let Some(tree_node) = self.tree.dependencies_tree.get(node_id) else { continue };
-            let dep_path = self.final_dep_path_of(node_id, final_dep_paths);
-            min_depth
-                .entry(dep_path)
-                .and_modify(|depth| *depth = (*depth).min(tree_node.depth))
-                .or_insert(tree_node.depth);
-        }
-
-        let mut record_dep_paths: HashMap<NodeId, DepPath> = HashMap::default();
-        let mut transitive_peer_dependencies_by_dep_path: HashMap<DepPath, HashSet<String>> =
-            HashMap::default();
-        for (node_id, record) in &self.node_records {
-            let dep_path = self.final_dep_path_of(node_id, final_dep_paths);
-            transitive_peer_dependencies_by_dep_path
-                .entry(dep_path.clone())
-                .or_default()
-                .extend(record.transitive_peer_dependencies.iter().cloned());
-            record_dep_paths.insert(node_id.clone(), dep_path);
-        }
-
-        let mut graph = DependenciesGraph::default();
-        let mut graph_order: HashMap<DepPath, u64> = HashMap::default();
-        for (node_id, record) in &self.node_records {
-            let dep_path = record_dep_paths[node_id].clone();
-            let depth = min_depth.get(&dep_path).copied().unwrap_or(record.depth);
-            let pkg_id = self.tree.dependencies_tree[node_id].resolved_package_id.clone();
-            let pkg = &self.tree.packages[&pkg_id];
-            let mut children: BTreeMap<String, DepPath> = BTreeMap::new();
-            for (alias, edge_node_id) in &record.edges {
-                children
-                    .insert(alias.clone(), self.final_dep_path_of(edge_node_id, final_dep_paths));
-            }
-            let resolved_peer_names: HashSet<String> = self
-                .node_external_peers
-                .get(node_id)
-                .map(|peers| peers.keys().cloned().collect())
-                .unwrap_or_default();
-            let mut candidate = DependenciesGraphNode {
-                dep_path: dep_path.clone(),
-                resolved_package_id: pkg_id.clone(),
-                resolve_result: Arc::clone(&pkg.result),
-                children,
-                optional_children: record.optional_child_aliases.clone(),
-                peer_dependencies: pkg.peer_dependencies.clone(),
-                transitive_peer_dependencies: record.transitive_peer_dependencies.clone(),
-                resolved_peer_names,
-                depth,
-                installable: record.installable,
-                is_pure: record.is_pure,
-                optional: pkg.optional,
-            };
-            match graph.entry(dep_path.clone()) {
-                std::collections::hash_map::Entry::Vacant(entry) => {
-                    graph_order.insert(dep_path.clone(), record.order);
-                    entry.insert(candidate);
-                }
-                std::collections::hash_map::Entry::Occupied(mut entry) => {
-                    let existing = entry.get();
-                    let existing_order =
-                        graph_order.get(&dep_path).copied().unwrap_or(record.order);
-                    let replace = candidate.depth < existing.depth
-                        || (candidate.depth == existing.depth && record.order < existing_order);
-                    if replace {
-                        candidate
-                            .transitive_peer_dependencies
-                            .extend(existing.transitive_peer_dependencies.iter().cloned());
-                        candidate
-                            .optional_children
-                            .extend(existing.optional_children.iter().cloned());
-                        merge_preferred_child_edges(
-                            &mut candidate,
-                            existing.children.clone(),
-                            &transitive_peer_dependencies_by_dep_path,
-                        );
-                        graph_order.insert(dep_path.clone(), record.order);
-                        entry.insert(candidate);
-                    } else {
-                        let existing = entry.get_mut();
-                        existing
-                            .transitive_peer_dependencies
-                            .extend(candidate.transitive_peer_dependencies);
-                        existing.optional_children.extend(candidate.optional_children);
-                        merge_preferred_child_edges(
-                            existing,
-                            candidate.children,
-                            &transitive_peer_dependencies_by_dep_path,
-                        );
-                    }
-                }
-            }
-        }
-        graph
-    }
-
-    fn optional_child_aliases(
-        &self,
-        pkg_id: &str,
-        edges: &BTreeMap<String, NodeId>,
-    ) -> HashSet<String> {
-        self.tree
-            .children_by_id
-            .get(pkg_id)
-            .into_iter()
-            .flat_map(|children| children.iter())
-            .filter(|edge| edge.optional && edges.contains_key(&edge.alias))
-            .map(|edge| edge.alias.clone())
-            .collect()
-    }
-
-    fn preview_peer_provider_children(
-        &mut self,
-        node_id: &NodeId,
-    ) -> (BTreeMap<String, NodeId>, Option<UndoRealize>) {
-        let (parent_ids, pkg_id, depth) = {
-            let node = &self.tree.dependencies_tree[node_id];
-            match &node.children {
-                TreeChildren::Realized(children) => {
-                    let providers = children
-                        .iter()
-                        .filter(|(alias, child_node_id)| {
-                            self.tree
-                                .dependencies_tree
-                                .get(*child_node_id)
-                                .and_then(|child| {
-                                    self.tree.packages.get(&child.resolved_package_id)
-                                })
-                                .is_some_and(|pkg| self.is_peer_relevant(alias, pkg))
-                        })
-                        .map(|(alias, child_node_id)| (alias.clone(), child_node_id.clone()))
-                        .collect();
-                    return (providers, None);
-                }
-                TreeChildren::Lazy { parent_ids } => {
-                    (parent_ids.clone(), node.resolved_package_id.clone(), node.depth)
-                }
-            }
-        };
-        let children = self.tree.children_by_id.get(&pkg_id).cloned().unwrap_or_default();
-        let provider_edge_indices = self
-            .peer_provider_children_by_pkg_id
-            .get(&pkg_id)
-            .map_or(&[][..], |providers| providers.relevant_edge_indices.as_slice());
-        let full_chain = parent_ids.pushed(pkg_id.clone());
-        let mut providers = BTreeMap::new();
-        let mut newly_inserted = Vec::new();
-        for &edge_index in provider_edge_indices {
-            let edge = &children[edge_index];
-            let Some(pkg) = self.tree.packages.get(&edge.pkg_id) else { continue };
-            if pkg_id == edge.pkg_id || full_chain.forms_cycle(&pkg_id, &edge.pkg_id) {
-                continue;
-            }
-            let child_node_id =
-                if pkg.is_leaf { NodeId::leaf(&edge.pkg_id) } else { NodeId::next() };
-            if !self.tree.dependencies_tree.contains_key(&child_node_id) {
-                self.tree.dependencies_tree.insert(
-                    child_node_id.clone(),
-                    DependenciesTreeNode::new(
-                        edge.pkg_id.clone(),
-                        TreeChildren::Lazy { parent_ids: full_chain.clone() },
-                        depth + 1,
-                        true,
-                    ),
-                );
-                newly_inserted.push(child_node_id.clone());
-            }
-            providers.insert(edge.alias.clone(), child_node_id);
-        }
-        (providers, Some(UndoRealize { newly_inserted, prev_parent_ids: parent_ids }))
-    }
-
-    /// Realize the `(alias → NodeId)` children of `node_id` if it's
-    /// currently a [`TreeChildren::Lazy`] entry; return the realized
-    /// map (cloned for the caller). On a [`TreeChildren::Realized`]
-    /// entry, just clones and returns. Expands the thunk on demand:
-    ///
-    /// 1. Walk [`ResolvedTree::children_by_id`] for this node's
-    ///    package id.
-    /// 2. Skip any child whose pkg id appears in `parent_ids` — that
-    ///    edge would form a cycle.
-    /// 3. For each surviving child, allocate a per-occurrence
-    ///    `NodeId` (leaves reuse the deterministic `NodeId::leaf`
-    ///    for the leaf-collapse the eager walker does too) and
-    ///    insert a fresh `dependencies_tree` entry with another
-    ///    `Lazy` children variant that carries `parent_ids +
-    ///    [self_pkg_id]` for cycle break on its own descendants.
-    /// 4. Flip this node's `children` field to `Realized` so a
-    ///    later visitor reuses the map.
-    fn realize_children(
-        &mut self,
-        node_id: &NodeId,
-    ) -> (BTreeMap<String, NodeId>, Option<UndoRealize>) {
-        self.realize_children_with(node_id, None)
-    }
-
-    fn realize_children_with(
-        &mut self,
-        node_id: &NodeId,
-        previewed: Option<&BTreeMap<String, NodeId>>,
-    ) -> (BTreeMap<String, NodeId>, Option<UndoRealize>) {
-        // Snapshot the bits we need; we'll mutate `self.tree` below
-        // and can't hold a borrow on the entry across the mutation.
-        let (parent_ids, pkg_id, depth) = {
-            let node = &self.tree.dependencies_tree[node_id];
-            match &node.children {
-                TreeChildren::Realized(map) => return (map.clone(), None),
-                TreeChildren::Lazy { parent_ids } => {
-                    (parent_ids.clone(), node.resolved_package_id.clone(), node.depth)
-                }
-            }
-        };
-        let children_spec = match self.tree.children_by_id.get(&pkg_id) {
-            Some(spec) => Arc::clone(spec),
-            // No spec means the first walk never recorded children
-            // for this package id — defensive empty case.
-            None => Arc::new(Vec::new()),
-        };
-        let child_depth = depth + 1;
-        let mut realized: BTreeMap<String, NodeId> = BTreeMap::new();
-        let mut newly_inserted: Vec<NodeId> = Vec::new();
-        let full_chain = parent_ids.pushed(pkg_id.clone());
-        for edge in children_spec.iter() {
-            // Same cycle gate as the eager walk: keep the first
-            // re-entry, drop a direct self-edge or a second lap.
-            if full_chain.forms_cycle(&pkg_id, &edge.pkg_id) {
-                continue;
-            }
-            // Reuse the first walk's classification (persisted on
-            // `ResolvedPackage::is_leaf` by `pkg_is_leaf`). Defaults
-            // to non-leaf when the package isn't in `packages` — same
-            // shape as the eager walker's `manifest == None` arm,
-            // and `NodeId::next()` keeps occurrences distinct so a
-            // later visit can still observe per-call-site state.
-            let previewed_node_id = previewed.and_then(|previewed| previewed.get(&edge.alias));
-            let child_node_id = if let Some(previewed_node_id) = previewed_node_id {
-                previewed_node_id.clone()
-            } else {
-                let is_leaf = self.tree.packages.get(&edge.pkg_id).is_some_and(|pkg| pkg.is_leaf);
-                if is_leaf { NodeId::leaf(&edge.pkg_id) } else { NodeId::next() }
-            };
-            let child_parent_ids = full_chain.clone();
-            if let Some(node) = self.tree.dependencies_tree.get_mut(&child_node_id) {
-                if node.depth > child_depth {
-                    node.depth = child_depth;
-                }
-            } else {
-                self.tree.dependencies_tree.insert(
-                    child_node_id.clone(),
-                    DependenciesTreeNode::new(
-                        edge.pkg_id.clone(),
-                        TreeChildren::Lazy { parent_ids: child_parent_ids },
-                        child_depth,
-                        true,
-                    ),
-                );
-                newly_inserted.push(child_node_id.clone());
-            }
-            realized.insert(edge.alias.clone(), child_node_id);
-        }
-        // Replace this node's `Lazy` with `Realized` so future
-        // visitors reuse the work.
-        if let Some(node) = self.tree.dependencies_tree.get_mut(node_id) {
-            node.children = TreeChildren::Realized(realized.clone());
-        }
-        (realized, Some(UndoRealize { newly_inserted, prev_parent_ids: parent_ids }))
-    }
-
-    fn previously_resolved_children(
-        &mut self,
-        parent_node_ids: &SharedChain<NodeId>,
-        parent_pkg_ids_chain: &SharedChain<String>,
-        current_pkg_id: &str,
-    ) -> BTreeMap<String, NodeId> {
-        let mut children = BTreeMap::new();
-        if !parent_pkg_ids_chain.iter().any(|pkg_id| pkg_id == current_pkg_id) {
-            return children;
-        }
-        for parent_node_id in parent_node_ids.iter() {
-            let same_pkg = self
-                .tree
-                .dependencies_tree
-                .get(parent_node_id)
-                .is_some_and(|node| node.resolved_package_id == current_pkg_id);
-            if same_pkg {
-                for (alias, child_node_id) in self.realize_children(parent_node_id).0 {
-                    children.entry(alias).or_insert(child_node_id);
-                }
-            }
-        }
-        children
-    }
-
-    fn add_graph_child_or_pending(
-        &mut self,
-        graph_children: &mut BTreeMap<String, DepPath>,
-        parent_dep_path: &DepPath,
-        alias: String,
-        node_id: NodeId,
-    ) {
-        if let Some(dep_path) = self.node_dep_paths.get(&node_id) {
-            graph_children.insert(alias, dep_path.clone());
-        } else if let Some(link_dep_path) = link_node_id_as_dep_path(&node_id) {
-            // `topParents` linked-dep NodeIds never enter the tree, so
-            // `node_dep_paths` is empty for them; the `link:<rel>`
-            // NodeId is itself a valid DepPath, so the snapshot's child
-            // edge can use it verbatim.
-            graph_children.insert(alias, link_dep_path);
-        } else {
-            self.pending_peer_edges.push(PendingPeerEdge {
-                parent_dep_path: parent_dep_path.clone(),
-                child_alias: alias,
-                child_node_id: node_id,
-            });
-        }
-    }
-
-    fn parent_refs_match(&self, current: &ParentRef, new: &ParentRef) -> bool {
-        if current.version != new.version || current.alias != new.alias {
-            return false;
-        }
-        let Some(current_name) = self.parent_ref_package_name(current) else {
-            return true;
-        };
-        let Some(new_name) = self.parent_ref_package_name(new) else {
-            return true;
-        };
-        current_name == new_name
-    }
-
-    fn parent_ref_package_name(&self, parent_ref: &ParentRef) -> Option<String> {
-        let node_id = parent_ref.node_id.as_ref()?;
-        let tree_node = self.tree.dependencies_tree.get(node_id)?;
-        let pkg = self.tree.packages.get(&tree_node.resolved_package_id)?;
-        Some(pkg_name_version(&pkg.result).0)
-    }
-
-    fn inherited_parent_pkg_breaks_peer_diamond(
-        &self,
-        parent_refs: &ParentRefs,
-        inherited_parent_pkg: &ParentRef,
-        own_child_parent_pkg: &ParentRef,
-        node_id: &NodeId,
-    ) -> bool {
-        let (Some(inherited_node_id), Some(own_child_node_id)) =
-            (inherited_parent_pkg.node_id.as_ref(), own_child_parent_pkg.node_id.as_ref())
-        else {
-            return false;
-        };
-        if inherited_node_id == own_child_node_id {
-            return false;
-        }
-        let Some(inherited_context) = self.parent_pkgs_of_node.get(inherited_node_id) else {
-            return false;
-        };
-        let Some(parent_pkg) = self
-            .tree
-            .dependencies_tree
-            .get(own_child_node_id)
-            .and_then(|node| self.tree.packages.get(&node.resolved_package_id))
-        else {
-            return false;
-        };
-        let (parent_pkg_name, _) = pkg_name_version(&parent_pkg.result);
-
-        let mut conflicting_peers = HashSet::default();
-        for peer_name in parent_pkg.peer_dependencies.keys() {
-            if !self.tree.all_peer_dep_names.contains(peer_name) {
-                continue;
-            }
-            let Some(inherited_peer) = inherited_context.get(peer_name) else { continue };
-            let Some(current_peer) = parent_refs.get(peer_name) else { continue };
-            if self.parent_peer_differs(current_peer, inherited_peer) {
-                conflicting_peers.insert(peer_name.clone());
-            }
-        }
-        if conflicting_peers.is_empty() {
-            return false;
-        }
-
-        let Some(node) = self.tree.dependencies_tree.get(node_id) else { return false };
-        let child_pkg_ids: Vec<&str> = match &node.children {
-            TreeChildren::Realized(children) => children
-                .values()
-                .filter_map(|child_node_id| self.tree.dependencies_tree.get(child_node_id))
-                .map(|child| child.resolved_package_id.as_str())
-                .collect(),
-            TreeChildren::Lazy { .. } => self
-                .tree
-                .children_by_id
-                .get(&node.resolved_package_id)
-                .into_iter()
-                .flat_map(|children| children.iter())
-                .map(|child| child.pkg_id.as_str())
-                .collect(),
-        };
-        for child_pkg_id in child_pkg_ids {
-            let Some(child_pkg) = self.tree.packages.get(child_pkg_id) else { continue };
-            if !child_pkg.peer_dependencies.contains_key(&parent_pkg_name) {
-                continue;
-            }
-            if conflicting_peers.iter().any(|peer| child_pkg.peer_dependencies.contains_key(peer)) {
-                return true;
-            }
-        }
-        false
-    }
-
-    fn parent_peer_differs(
-        &self,
-        current_peer: &ParentRef,
-        inherited_peer: &ParentPkgInfo,
-    ) -> bool {
-        if let Some(inherited_pkg_id) = inherited_peer.pkg_id.as_ref() {
-            let Some(current_node_id) = current_peer.node_id.as_ref() else {
-                return true;
-            };
-            return self
-                .tree
-                .dependencies_tree
-                .get(current_node_id)
-                .is_none_or(|node| node.resolved_package_id != *inherited_pkg_id);
-        }
-        inherited_peer.version.as_ref() != Some(&current_peer.version)
-    }
-
-    /// Build the `(peer_name → ParentPkgInfo)` snapshot that gets
-    /// stored on [`Self::parent_pkgs_of_node`] for each child the
-    /// caller is about to descend into.
-    ///
-    /// `link:` parents don't have a real tree entry; pacquet's
-    /// [`ParentRef`] keeps the `NodeId` but the tree-lookup falls back
-    /// to a pure `version` comparison.
-    fn parent_dep_paths_from_refs(
-        &self,
-        parent_refs: &ParentRefs,
-    ) -> Arc<HashMap<String, ParentPkgInfo>> {
-        let mut out = HashMap::default();
-        for (name, parent_ref) in parent_refs {
-            if !self.tree.all_peer_dep_names.contains(name) {
-                continue;
-            }
-            let pkg_id = parent_ref
-                .node_id
-                .as_ref()
-                .and_then(|nid| self.tree.dependencies_tree.get(nid))
-                .map(|tn| tn.resolved_package_id.clone());
-            let version = pkg_id.is_none().then(|| parent_ref.version.clone());
-            out.insert(
-                name.clone(),
-                ParentPkgInfo {
-                    pkg_id,
-                    version,
-                    depth: parent_ref.depth,
-                    occurrence: parent_ref.occurrence,
-                },
-            );
-        }
-        // Shared, not cloned: the same snapshot is recorded for every
-        // child of a node, and these maps dominated the walker's
-        // allocation churn when cloned per child.
-        Arc::new(out)
-    }
-
-    fn remember_parent_context_if_peer_provider(
-        &mut self,
-        alias: &str,
-        node_id: &NodeId,
-        parent_context: &Arc<HashMap<String, ParentPkgInfo>>,
-    ) {
-        let Some(tree_node) = self.tree.dependencies_tree.get(node_id) else { return };
-        let Some(pkg) = self.tree.packages.get(&tree_node.resolved_package_id) else { return };
-        if self.is_peer_relevant(alias, pkg) {
-            self.parent_pkgs_of_node.insert(node_id.clone(), Arc::clone(parent_context));
-        }
-    }
-
-    fn owned_package(&mut self, pkg_id: &str) -> Arc<ResolvedPackage> {
-        if let Some(pkg) = self.packages_by_id.get(pkg_id) {
-            return Arc::clone(pkg);
-        }
-        let pkg = Arc::new(self.tree.packages[pkg_id].clone());
-        self.packages_by_id.insert(pkg_id.to_string(), Arc::clone(&pkg));
-        pkg
-    }
-
-    fn remember_resolved_node(&mut self, node_id: &NodeId, dep_path: &DepPath) {
-        let retain = !self.discovery
-            || self.parent_pkgs_of_node.contains_key(node_id)
-            || self.opts.hoisted_peer_provider_node_ids.contains(node_id);
-        if !retain {
-            return;
-        }
-        self.node_dep_paths.insert(node_id.clone(), dep_path.clone());
-        self.visited_this_call.insert(node_id.clone());
-    }
-
-    fn undo_realize(
-        &mut self,
-        node_id: &NodeId,
-        undo: Option<UndoRealize>,
-        output: Option<&NodeOutput>,
-    ) {
-        let Some(undo) = undo else { return };
-        for child_id in &undo.newly_inserted {
-            if should_retain_materialized_node(&self.retained_peer_node_ids, output, child_id) {
-                continue;
-            }
-            self.tree.dependencies_tree.remove(child_id);
-            self.parent_pkgs_of_node.remove(child_id);
-            self.node_dep_paths.remove(child_id);
-            self.visited_this_call.remove(child_id);
-        }
-        if let Some(node) = self.tree.dependencies_tree.get_mut(node_id) {
-            node.children = TreeChildren::Lazy { parent_ids: undo.prev_parent_ids };
-        }
-    }
-
-    fn find_fast_hit(
-        &self,
-        node_id: &NodeId,
-        parent_refs: &ParentRefs,
-        pkg_id: &str,
-    ) -> Option<&PeersCacheItem> {
-        let TreeChildren::Lazy { parent_ids } = &self.tree.dependencies_tree.get(node_id)?.children
-        else {
-            return None;
-        };
-        self.find_fast_hit_for_lazy(parent_ids, parent_refs, pkg_id)
-    }
-
-    fn find_fast_hit_for_lazy(
-        &self,
-        parent_ids: &AncestorIds,
-        parent_refs: &ParentRefs,
-        pkg_id: &str,
-    ) -> Option<&PeersCacheItem> {
-        self.peers_cache.get(pkg_id)?.iter().find(|item| {
-            matches!(
-                self.fast_cache_item_matches(parent_ids, parent_refs, pkg_id, item),
-                FastCacheMatch::Match,
-            )
-        })
-    }
-
-    fn deferred_child_resolution(
-        &self,
-        parent_ids: &AncestorIds,
-        parent_refs: &ParentRefs,
-        pkg_id: &str,
-    ) -> DeferredChildResolution {
-        if self
-            .pure_pkgs
-            .get(pkg_id)
-            .is_some_and(|_| self.tree.packages[pkg_id].peer_dependencies.is_empty())
-        {
-            return DeferredChildResolution::Pure(self.pure_pkgs[pkg_id].clone());
-        }
-        if let Some(cached) = self
-            .find_fast_hit_for_lazy(parent_ids, parent_refs, pkg_id)
-            .map(PeersCacheItem::to_cached_node_output)
-            && cached.output.missing_peers.is_empty()
-        {
-            return DeferredChildResolution::Cached(cached);
-        }
-        DeferredChildResolution::Materialize(pkg_id.to_string())
-    }
-
-    fn resolve_deferred_child(&mut self, context: DeferredChildContext<'_>) -> NodeOutput {
-        let DeferredChildContext {
-            edge,
-            node_id,
-            parent_ids,
-            parent_refs,
-            parent_dep_paths,
-            chain_names,
-            parent_node_ids,
-            parent_pkg_ids,
-            depth,
-        } = context;
-        match self.deferred_child_resolution(parent_ids, parent_refs, &edge.pkg_id) {
-            DeferredChildResolution::Pure(dep_path) => NodeOutput {
-                dep_path,
-                external_resolved_peers: Arc::clone(&self.empty_resolved_peers),
-                auto_install_resolved_peers: HashMap::default(),
-                missing_peers: Arc::clone(&self.empty_missing_peers),
-                subtree_missing_by_pkg: None,
-            },
-            DeferredChildResolution::Cached(cached) => cached.output,
-            DeferredChildResolution::Materialize(pkg_id) => {
-                self.tree.dependencies_tree.insert(
-                    node_id.clone(),
-                    DependenciesTreeNode::new(
-                        pkg_id,
-                        TreeChildren::Lazy { parent_ids: parent_ids.clone() },
-                        depth,
-                        true,
-                    ),
-                );
-                let output = self.resolve_node(
-                    &node_id,
-                    parent_refs,
-                    parent_dep_paths,
-                    chain_names,
-                    parent_node_ids,
-                    parent_pkg_ids,
-                );
-                if !self.parent_pkgs_of_node.contains_key(&node_id)
-                    && !should_retain_materialized_node(
-                        &self.retained_peer_node_ids,
-                        Some(&output),
-                        &node_id,
-                    )
-                {
-                    self.tree.dependencies_tree.remove(&node_id);
-                    self.node_dep_paths.remove(&node_id);
-                    self.visited_this_call.remove(&node_id);
-                }
-                output
-            }
-        }
-    }
-
-    fn fast_cache_item_matches(
-        &self,
-        parent_ids: &AncestorIds,
-        parent_refs: &ParentRefs,
-        pkg_id: &str,
-        item: &PeersCacheItem,
-    ) -> FastCacheMatch {
-        let mut ambiguous = false;
-        for (name, cached_node_id) in item.resolved_peers.iter() {
-            match self.fast_provider_for_name(parent_ids, parent_refs, pkg_id, name) {
-                FastProvider::Missing => return FastCacheMatch::NoMatch,
-                FastProvider::Inherited(current_ref) => {
-                    if !self.parent_ref_matches_cached(current_ref, cached_node_id) {
-                        return FastCacheMatch::NoMatch;
-                    }
-                }
-                FastProvider::Child(child_pkg_id) => {
-                    let Some(cached_tree_node) = self.tree.dependencies_tree.get(cached_node_id)
-                    else {
-                        return FastCacheMatch::NoMatch;
-                    };
-                    if cached_tree_node.resolved_package_id != child_pkg_id {
-                        return FastCacheMatch::NoMatch;
-                    }
-                    let child_is_stable = self.pure_pkgs.contains_key(child_pkg_id)
-                        || matches!(
-                            cached_node_id,
-                            NodeId::Leaf(cached_pkg_id) if cached_pkg_id.as_ref() == child_pkg_id,
-                        );
-                    if !child_is_stable {
-                        ambiguous = true;
-                    }
-                }
-                FastProvider::Ambiguous => ambiguous = true,
-            }
-        }
-        for name in item.missing_peers.keys() {
-            match self.fast_provider_for_name(parent_ids, parent_refs, pkg_id, name) {
-                FastProvider::Missing => {}
-                FastProvider::Inherited(_) | FastProvider::Child(_) => {
-                    return FastCacheMatch::NoMatch;
-                }
-                FastProvider::Ambiguous => ambiguous = true,
-            }
-        }
-        if ambiguous { FastCacheMatch::Ambiguous } else { FastCacheMatch::Match }
-    }
-
-    fn fast_provider_for_name<'a>(
-        &'a self,
-        parent_ids: &AncestorIds,
-        parent_refs: &'a ParentRefs,
-        pkg_id: &str,
-        name: &str,
-    ) -> FastProvider<'a> {
-        let inherited = parent_refs.get(name);
-        let Some(children) = self.tree.children_by_id.get(pkg_id) else {
-            return inherited.map_or(FastProvider::Missing, FastProvider::Inherited);
-        };
-        let Some(edge_indices) = self
-            .peer_provider_children_by_pkg_id
-            .get(pkg_id)
-            .and_then(|providers| providers.edge_indices_by_name.get(name))
-        else {
-            return inherited.map_or(FastProvider::Missing, FastProvider::Inherited);
-        };
-
-        let mut child_pkg_id = None;
-        for &edge_index in edge_indices {
-            let edge = &children[edge_index];
-            if parent_ids.forms_cycle(pkg_id, &edge.pkg_id) {
-                continue;
-            }
-            if child_pkg_id.is_some() {
-                return FastProvider::Ambiguous;
-            }
-            child_pkg_id = Some(edge.pkg_id.as_str());
-        }
-
-        match (inherited, child_pkg_id) {
-            (Some(_), Some(_)) => FastProvider::Ambiguous,
-            (Some(parent_ref), None) => FastProvider::Inherited(parent_ref),
-            (None, Some(pkg_id)) => FastProvider::Child(pkg_id),
-            (None, None) => FastProvider::Missing,
-        }
-    }
-
-    fn parent_ref_matches_cached(&self, current_ref: &ParentRef, cached_node_id: &NodeId) -> bool {
-        let Some(current_node_id) = current_ref.node_id.as_ref() else {
-            return false;
-        };
-        if current_node_id == cached_node_id {
-            return true;
-        }
-        if let (Some(cached_dp), Some(current_dp)) =
-            (self.node_dep_paths.get(cached_node_id), self.node_dep_paths.get(current_node_id))
-            && cached_dp == current_dp
-        {
-            return true;
-        }
-        let (Some(cached_tree_node), Some(current_tree_node)) = (
-            self.tree.dependencies_tree.get(cached_node_id),
-            self.tree.dependencies_tree.get(current_node_id),
-        ) else {
-            return false;
-        };
-        let parent_pkg_id = &current_tree_node.resolved_package_id;
-        if parent_pkg_id != &cached_tree_node.resolved_package_id {
-            return false;
-        }
-        self.pure_pkgs.contains_key(parent_pkg_id)
-            || self.parent_packages_match(cached_node_id, current_node_id)
-    }
-
-    /// Look up [`Self::peers_cache`] for a cached resolution of
-    /// `pkg_id` whose parent peer context is compatible with the
-    /// current `parent_refs`.
-    ///
-    /// A cache item matches when, for every cached resolved peer:
-    ///
-    /// 1. The current `parent_refs` has a counterpart entry for the
-    ///    same name with a real `NodeId`.
-    /// 2. Either the two `NodeId`s are equal, OR they map to the
-    ///    same already-computed [`DepPath`] in
-    ///    [`Self::node_dep_paths`], OR the two tree-nodes' resolved
-    ///    package ids match — and in the package-id match case, the
-    ///    deep [`Self::parent_packages_match`] check on the two
-    ///    parents' own recorded contexts also succeeds (unless the
-    ///    package id is itself in [`Self::pure_pkgs`], which makes
-    ///    the deep check vacuous).
-    /// 3. None of the cache item's missing-peer names are satisfied
-    ///    by the current `parent_refs` — a name the cache walk
-    ///    recorded as missing must still be missing here.
-    fn find_hit(&self, parent_refs: &ParentRefs, pkg_id: &str) -> Option<&PeersCacheItem> {
-        let cache_items = self.peers_cache.get(pkg_id)?;
-        cache_items.iter().find(|item| {
-            for (name, cached_node_id) in item.resolved_peers.iter() {
-                let Some(current_ref) = parent_refs.get(name) else {
-                    return false;
-                };
-                if !self.parent_ref_matches_cached(current_ref, cached_node_id) {
-                    return false;
-                }
-            }
-            for missing_name in item.missing_peers.keys() {
-                if parent_refs.contains_key(missing_name) {
-                    return false;
-                }
-            }
-            true
-        })
-    }
-
-    /// Compare two `NodeId`s' recorded parent peer contexts:
-    /// both nodes' contexts must have the same set of peer-relevant
-    /// names, every name must resolve to the same version or
-    /// `pkgIdWithPatchHash`, and — when a peer is shadowed (an
-    /// `occurrence > 0` somewhere on either side) — the contexts
-    /// must additionally agree on depth/`purePkgs` to compensate
-    /// for the loss of single-occurrence guarantees on the
-    /// shallow-equality path.
-    fn parent_packages_match(&self, cached_node_id: &NodeId, current_node_id: &NodeId) -> bool {
-        let Some(cached_parents) = self.parent_pkgs_of_node.get(cached_node_id) else {
-            return false;
-        };
-        let Some(current_parents) = self.parent_pkgs_of_node.get(current_node_id) else {
-            return false;
-        };
-        if cached_parents.len() != current_parents.len() {
-            return false;
-        }
-        let max_depth = current_parents.values().map(|info| info.depth).max().unwrap_or(0);
-        let peer_deps_not_shadowed = parent_pkgs_have_single_occurrence(cached_parents)
-            && parent_pkgs_have_single_occurrence(current_parents);
-        for (name, cached_info) in cached_parents.iter() {
-            let Some(current_info) = current_parents.get(name) else { return false };
-            // Version-only match covers `link:` parents only when
-            // both recorded contexts are version-only.
-            if let (Some(cached_version), Some(current_version)) =
-                (&cached_info.version, &current_info.version)
-            {
-                if cached_version == current_version {
-                    continue;
-                }
-                return false;
-            }
-            // Package-id match with shadowing guard.
-            let Some(cached_pkg_id) = cached_info.pkg_id.as_ref() else { return false };
-            if cached_info.pkg_id != current_info.pkg_id {
-                return false;
-            }
-            if !(peer_deps_not_shadowed
-                || current_info.depth == max_depth
-                || self.pure_pkgs.contains_key(cached_pkg_id))
-            {
-                return false;
-            }
-        }
-        true
-    }
-}
-
-fn should_retain_materialized_node(
-    retained_peer_node_ids: &HashSet<NodeId>,
-    output: Option<&NodeOutput>,
-    node_id: &NodeId,
-) -> bool {
-    retained_peer_node_ids.contains(node_id)
-        || output.is_some_and(|output| {
-            output.external_resolved_peers.values().any(|resolved_id| resolved_id == node_id)
-                || output
-                    .auto_install_resolved_peers
-                    .values()
-                    .any(|resolved_id| resolved_id == node_id)
-        })
-}
-
-/// Whether every entry in `parents` has `occurrence == 0`.
-fn parent_pkgs_have_single_occurrence(parents: &HashMap<String, ParentPkgInfo>) -> bool {
-    parents.values().all(|info| info.occurrence == 0)
-}
-
-fn merge_missing_summary(
-    target: &mut HashMap<String, HashSet<String>>,
-    seen: &mut HashSet<usize>,
-    summary: SubtreeMissingByPkg,
-) {
-    let Some(summary) = summary else { return };
-    let mut pending = vec![summary];
-    while let Some(summary) = pending.pop() {
-        if !seen.insert(Arc::as_ptr(&summary) as usize) {
-            continue;
-        }
-        if let Some((pkg_id, names)) = &summary.own {
-            target.entry(pkg_id.clone()).or_default().extend(names.iter().cloned());
-        }
-        pending.extend(summary.children.iter().cloned());
-    }
-}
-
-fn merge_preferred_child_edges(
-    target: &mut DependenciesGraphNode,
-    children: BTreeMap<String, DepPath>,
-    transitive_peer_dependencies_by_dep_path: &HashMap<DepPath, HashSet<String>>,
-) {
-    let available_peer_names =
-        available_peer_names_for_dep_path(&target.dep_path, &target.resolve_result);
-    for (alias, candidate_dep_path) in children {
-        match target.children.entry(alias) {
-            std::collections::btree_map::Entry::Vacant(entry) => {
-                entry.insert(candidate_dep_path);
-            }
-            std::collections::btree_map::Entry::Occupied(mut entry) => {
-                let preferred = child_dep_path_is_preferred(
-                    entry.get(),
-                    &candidate_dep_path,
-                    &available_peer_names,
-                    transitive_peer_dependencies_by_dep_path,
-                );
-                if preferred {
-                    entry.insert(candidate_dep_path);
-                }
-            }
-        }
-    }
-}
-
-fn available_peer_names_for_dep_path(
-    dep_path: &DepPath,
-    resolve_result: &ResolveResult,
-) -> HashSet<String> {
-    let mut names: HashSet<String> =
-        peer_segment_names(dep_path).unwrap_or_default().into_iter().collect();
-    names.insert(pkg_name_version(resolve_result).0);
-    names
-}
-
-fn child_dep_path_is_preferred(
-    current: &DepPath,
-    candidate: &DepPath,
-    available_peer_names: &HashSet<String>,
-    transitive_peer_dependencies_by_dep_path: &HashMap<DepPath, HashSet<String>>,
-) -> bool {
-    if current == candidate {
-        return false;
-    }
-    let current_unavailable = unavailable_non_transitive_peer_segment_names(
-        current,
-        available_peer_names,
-        transitive_peer_dependencies_by_dep_path,
-    )
-    .unwrap_or_default();
-    let candidate_unavailable = unavailable_non_transitive_peer_segment_names(
-        candidate,
-        available_peer_names,
-        transitive_peer_dependencies_by_dep_path,
-    )
-    .unwrap_or_default();
-    if candidate_unavailable.len() != current_unavailable.len() {
-        return candidate_unavailable.len() < current_unavailable.len();
-    }
-    if !candidate_unavailable.is_empty() {
-        return false;
-    }
-    let current_peer_count =
-        available_peer_segment_count(current, available_peer_names).unwrap_or(0);
-    let candidate_peer_count =
-        available_peer_segment_count(candidate, available_peer_names).unwrap_or(0);
-    candidate_peer_count > current_peer_count
-}
-
-fn available_peer_segment_count(
-    dep_path: &DepPath,
-    available_peer_names: &HashSet<String>,
-) -> Option<usize> {
-    let names = peer_segment_names(dep_path)?;
-    Some(names.into_iter().filter(|name| available_peer_names.contains(name)).count())
-}
-
-fn unavailable_non_transitive_peer_segment_names(
-    dep_path: &DepPath,
-    available_peer_names: &HashSet<String>,
-    transitive_peer_dependencies_by_dep_path: &HashMap<DepPath, HashSet<String>>,
-) -> Option<Vec<String>> {
-    let transitive_peer_dependencies = transitive_peer_dependencies_by_dep_path.get(dep_path);
-    let names = peer_segment_names(dep_path)?;
-    Some(
-        names
-            .into_iter()
-            .filter(|name| {
-                !available_peer_names.contains(name)
-                    && transitive_peer_dependencies
-                        .is_none_or(|transitive| !transitive.contains(name))
-            })
+            .map(|dep| (dep.alias.clone(), dep.node_id.clone()))
             .collect(),
-    )
-}
-
-fn peer_segment_names(dep_path: &DepPath) -> Option<Vec<String>> {
-    let raw = dep_path.as_str();
-    let suffix = index_of_dep_path_suffix(raw);
-    let peers_index = suffix.peers_index?;
-    let segments = split_peer_suffix_segments(&raw[peers_index..])?;
-    segments.iter().map(|segment| peer_segment_name(segment).map(str::to_string)).collect()
-}
-
-fn split_peer_suffix_segments(suffix: &str) -> Option<Vec<String>> {
-    let bytes = suffix.as_bytes();
-    let mut segments = Vec::new();
-    let mut depth = 0i32;
-    let mut start = None;
-    for (idx, byte) in bytes.iter().enumerate() {
-        match byte {
-            b'(' => {
-                if depth == 0 {
-                    start = Some(idx + 1);
-                }
-                depth += 1;
-            }
-            b')' => {
-                depth -= 1;
-                if depth < 0 {
-                    return None;
-                }
-                if depth == 0 {
-                    let start = start.take()?;
-                    segments.push(suffix[start..idx].to_string());
-                }
-            }
-            _ if depth == 0 => return None,
-            _ => {}
-        }
-    }
-    (depth == 0).then_some(segments)
-}
-
-fn peer_segment_name(segment: &str) -> Option<&str> {
-    let version_separator = if let Some(unscoped) = segment.strip_prefix('@') {
-        1 + unscoped.find('@')?
-    } else {
-        segment.find('@')?
+        declared_direct_dependencies: HashSet::default(),
+        explicitly_requested_direct_dependencies: HashSet::default(),
     };
-    Some(&segment[..version_separator])
-}
-
-/// Record a parent ref under both keys: each parent is recorded by its
-/// install alias *and* its real name when the two differ.
-///
-/// `parent_node_id` is the [`NodeId`] the parent should appear under
-/// in the [`ParentRefs`] map. For most parents this is just
-/// `direct.node_id`, but `link:` parents may carry the remapped
-/// node id produced by [`remap_link_node_id`] when
-/// `excludeLinksFromLockfile` is on.
-fn insert_parent_ref(
-    refs: &mut ParentRefs,
-    direct_alias: &str,
-    parent_node_id: NodeId,
-    pkg: &ResolvedPackage,
-    depth: i32,
-) {
-    let (real_name, version) = pkg_name_version(&pkg.result);
-    let parent_ref = ParentRef {
-        version,
-        node_id: Some(parent_node_id),
-        alias: (direct_alias != real_name).then(|| direct_alias.to_string()),
-        depth,
-        occurrence: 0,
-    };
-    update_parent_refs(refs, direct_alias, &parent_ref);
-    if direct_alias != real_name {
-        update_parent_refs(refs, &real_name, &parent_ref);
+    let mut sources = vec![source_of(importer)];
+    if let Some(root) = root_importer
+        && root.id != importer.id
+    {
+        sources.push(source_of(root));
     }
+    sources
 }
 
-fn update_parent_refs(refs: &mut ParentRefs, new_alias: &str, parent_ref: &ParentRef) {
-    if let Some(existing) = refs.get(new_alias) {
-        let existing_has_alias = existing.alias.as_deref().is_some_and(|alias| alias != new_alias);
-        if !existing_has_alias {
-            return;
-        }
-        let new_has_alias = parent_ref.alias.as_deref().is_some_and(|alias| alias != new_alias);
-        if new_has_alias && version_gte(&existing.version, &parent_ref.version) {
-            return;
-        }
-    }
-    refs.insert(new_alias.to_string(), parent_ref.clone());
-}
-
-fn version_gte(left: &str, right: &str) -> bool {
-    match (Version::parse(left), Version::parse(right)) {
-        (Ok(left), Ok(right)) => left >= right,
-        _ => left >= right,
-    }
-}
-
-/// Compute the `link:` [`NodeId`] under which a workspace-link parent
-/// should appear in [`ParentRefs`] when
-/// [`ResolvePeersOptions::exclude_links_from_lockfile`] is on.
-///
-/// Returns `None` when:
-///
-/// - the dep isn't a `link:` directory resolution;
-/// - the setting is off or the lockfile / modules dirs are missing;
-/// - the link target lives under `lockfile_dir` (workspace-internal
-///   link — already stable across machines, no remap needed).
-///
-/// On `Some`, the remap encodes `<modules_dir>/<alias>` as a path
-/// relative to `lockfile_dir`, prefixed with `link:`.
-fn remap_link_node_id(
+/// The upstream `getNodeIdsByPreviousDepPath`: first node (in
+/// deterministic id order) claiming each `previous_dep_path`. Empty
+/// unless a locked-peer reuse pass was requested.
+fn build_node_ids_by_previous_dep_path(
+    tree: &ResolvedTree,
     opts: &ResolvePeersOptions,
-    alias: &str,
-    result: &ResolveResult,
-) -> Option<NodeId> {
-    if !opts.exclude_links_from_lockfile {
-        return None;
+) -> HashMap<DepPath, NodeId> {
+    let mut map = HashMap::default();
+    if opts.resolved_peer_provider_paths.is_none() {
+        return map;
     }
-    let lockfile_dir = opts.lockfile_dir.as_ref()?;
-    let modules_dir = opts.modules_dir.as_ref()?;
-    let directory = match &result.resolution {
-        pacquet_lockfile::LockfileResolution::Directory(dir) => &dir.directory,
-        _ => return None,
-    };
-    let link_target = std::path::Path::new(directory);
-    if pacquet_fs::is_subdir(lockfile_dir, link_target) {
-        return None;
+    let mut node_ids: Vec<&NodeId> = tree.dependencies_tree.keys().collect();
+    node_ids.sort_by_key(|node_id| node_id_sort_key(node_id));
+    for node_id in node_ids {
+        if let Some(previous) = tree.dependencies_tree[node_id].previous_dep_path.as_ref()
+            && !map.contains_key(previous)
+        {
+            map.insert(previous.clone(), node_id.clone());
+        }
     }
-    let target = modules_dir.join(alias);
-    let rel = pathdiff::diff_paths(&target, lockfile_dir)?;
-    let rel = rel.display().to_string().replace('\\', "/");
-    Some(NodeId::leaf(&format!("link:{rel}")))
+    map
 }
 
-/// Build the `parents` chain attached to a peer issue. Records just
-/// `name` and `version` per parent, which is what the renderer
-/// downstream consumes.
-fn parents_from_chain(chain_names: &SharedChain<String>, _pkg_name: &str) -> Vec<ParentPackageRef> {
-    // The chain pacquet tracks today is name-only — populating
-    // `version` would need a parallel `Vec<String>` of versions or a
-    // re-lookup against the tree. The issue-renderer consumes the
-    // names primarily; expanding to versions is a follow-up.
-    chain_names
-        .to_root_vec()
-        .into_iter()
-        .map(|name| ParentPackageRef { name, version: String::new() })
-        .collect()
-}
-
-/// Range-satisfaction check that tolerates prereleases the way
-/// `@yarnpkg/core/semverUtils.satisfiesWithPrereleases` does — falls
-/// back to a literal-equality check when the range can't be parsed,
-/// which lets non-semver peer ranges (`*`, git refs, etc.) still
-/// match.
-///
-/// **Prerelease tolerance.** `node-semver`'s [`Range::satisfies`]
-/// rejects prerelease versions against non-prerelease comparators.
-/// Yarn's `satisfiesWithPrereleases` explicitly allows that pairing.
-/// We approximate it by retrying with the prerelease tag stripped: if
-/// `version` is a prerelease and the straight check fails, see whether
-/// the base `MAJOR.MINOR.PATCH` satisfies the range — without pulling
-/// in Yarn's full per-comparator algorithm.
-fn satisfies_with_prereleases(version: &str, range: &str) -> bool {
-    if range == "*" {
-        return true;
-    }
-    let Ok(parsed_version) = Version::parse(version) else {
-        return version == range;
-    };
-    let Ok(parsed_range) = Range::parse(range) else {
-        return version == range;
-    };
-    if parsed_version.satisfies(&parsed_range) {
-        return true;
-    }
-    if !parsed_version.is_prerelease() {
-        return false;
-    }
-    let base = Version {
-        major: parsed_version.major,
-        minor: parsed_version.minor,
-        patch: parsed_version.patch,
-        pre_release: Vec::new(),
-        build: Vec::new(),
-    };
-    base.satisfies(&parsed_range)
-}
-
+#[cfg(test)]
+mod test_support;
 #[cfg(test)]
 mod tests;
