+++ pnpm/crates/resolving-deps-resolver/src/resolve_peers/tests.rs
@@ -1,143 +1,16 @@
 use super::{
-    ImporterPeerInput, NodeOutput, NodeRecord, ParentRef, PeerDiscoveryCaches, ResolvePeersOptions,
-    Walker, importer_relative_link_dep_path, peer_segment_names, resolve_peers,
-    resolve_peers_workspace, satisfies_with_prereleases, scoped_hoisted_optional_parent_refs,
-    should_retain_materialized_node,
+    ImporterPeerInput, ResolvePeersOptions, resolve_peers, resolve_peers_workspace,
+    test_support::{
+        linked_package, package, package_with_peer_dependencies, tree_node, walker_for_tests,
+    },
 };
 use crate::{
     node_id::NodeId,
-    resolved_tree::{
-        DependenciesTreeNode, DirectDep, PeerDep, ResolvedPackage, ResolvedTree, TreeChildren,
-    },
+    resolved_tree::{DirectDep, ResolvedTree},
 };
 use pacquet_deps_path::DepPath;
-use pacquet_lockfile::{
-    DirectoryResolution, LockfileResolution, PkgName, PkgNameVer, TarballResolution,
-};
-use pacquet_resolving_resolver_base::{PkgResolutionId, ResolveResult};
 use rustc_hash::{FxHashMap as HashMap, FxHashSet as HashSet};
-use std::{collections::BTreeMap, path::Path, str::FromStr, sync::Arc};
-
-const PATCHED_WORKFLOWS_SDK: &str = concat!(
-    "@medusajs/workflows-sdk@2.13.3",
-    "(patch_hash=248195172cff27c28650c005b6aa0aa3b2f2976f9739544b360b81668f2d8b59)",
-    "(@types/node@20.19.17)",
-    "(better-sqlite3@12.8.0)",
-    "(express@4.21.2)",
-);
-
-#[test]
-fn materialized_nodes_referenced_by_peer_outputs_are_retained() {
-    let referenced = NodeId::next();
-    let unreferenced = NodeId::next();
-    let output = NodeOutput {
-        dep_path: DepPath::from("consumer@1.0.0"),
-        external_resolved_peers: Arc::new(HashMap::from_iter([(
-            "peer".to_string(),
-            referenced.clone(),
-        )])),
-        auto_install_resolved_peers: HashMap::default(),
-        missing_peers: Arc::new(HashMap::default()),
-        subtree_missing_by_pkg: None,
-    };
-
-    assert!(should_retain_materialized_node(&HashSet::default(), Some(&output), &referenced));
-    assert!(!should_retain_materialized_node(&HashSet::default(), Some(&output), &unreferenced,));
-    assert!(should_retain_materialized_node(
-        &HashSet::from_iter([unreferenced.clone()]),
-        None,
-        &unreferenced,
-    ));
-}
-
-#[test]
-fn locked_direct_dep_only_sees_its_locked_hoisted_optional_peers() {
-    let debug = NodeId::next();
-    let supports_color = NodeId::next();
-    let regular = NodeId::next();
-    let parent = |version: &str, node_id| ParentRef {
-        version: version.to_string(),
-        node_id: Some(node_id),
-        alias: None,
-        depth: 0,
-        occurrence: 0,
-    };
-    let parent_refs = HashMap::from_iter([
-        ("debug".to_string(), parent("4.4.3", debug.clone())),
-        ("supports-color".to_string(), parent("8.1.1", supports_color.clone())),
-        ("regular".to_string(), parent("1.0.0", regular)),
-    ]);
-
-    let scoped = scoped_hoisted_optional_parent_refs(
-        &parent_refs,
-        &HashSet::from_iter(["supports-color".to_string()]),
-        &HashSet::from_iter([debug, supports_color]),
-    );
-
-    assert!(!scoped.contains_key("debug"));
-    assert!(scoped.contains_key("supports-color"));
-    assert!(scoped.contains_key("regular"));
-}
-
-#[test]
-fn importer_relative_self_link_keeps_an_empty_target() {
-    let workspace = Path::new("workspace");
-    assert_eq!(
-        importer_relative_link_dep_path(&DepPath::from("link:."), Some(workspace), Some(workspace),),
-        DepPath::from("link:"),
-    );
-}
-
-/// The link target is relative to the importer, so a project dir that
-/// still carries `.` / `..` segments must be normalized before it is used
-/// as the base — otherwise those segments are counted as real directories
-/// and the target gains extra `..` hops.
-#[test]
-fn importer_relative_link_normalizes_the_project_dir() {
-    let expected = DepPath::from("link:../lib");
-    for project_dir in
-        ["workspace/packages/app", "workspace/packages/./app", "workspace/packages/nested/../app"]
-    {
-        assert_eq!(
-            importer_relative_link_dep_path(
-                &DepPath::from("link:packages/lib"),
-                Some(Path::new("workspace")),
-                Some(Path::new(project_dir)),
-            ),
-            expected,
-            "unexpected link target for project dir {project_dir:?}",
-        );
-    }
-}
-
-#[test]
-fn parses_peer_suffix_after_patch_hash() {
-    let dep_path = DepPath::from(PATCHED_WORKFLOWS_SDK);
-    assert_eq!(
-        peer_segment_names(&dep_path),
-        Some(vec!["@types/node".to_string(), "better-sqlite3".to_string(), "express".to_string(),]),
-    );
-}
-
-#[test]
-fn satisfies_handles_basic_ranges() {
-    assert!(satisfies_with_prereleases("1.2.3", "^1.0.0"));
-    assert!(!satisfies_with_prereleases("2.0.0", "^1.0.0"));
-    assert!(satisfies_with_prereleases("18.0.0", "*"));
-}
-
-#[test]
-fn satisfies_falls_back_to_equality_for_unparsable_ranges() {
-    assert!(satisfies_with_prereleases("workspace:^1.0.0", "workspace:^1.0.0"));
-    assert!(!satisfies_with_prereleases("1.0.0", "workspace:^1.0.0"));
-}
-
-#[test]
-fn satisfies_accepts_prerelease_against_non_prerelease_range() {
-    assert!(satisfies_with_prereleases("18.0.0-rc.1", "^18.0.0"));
-    assert!(satisfies_with_prereleases("1.2.3-beta.0", "^1.2.0"));
-    assert!(!satisfies_with_prereleases("19.0.0-rc.1", "^18.0.0"));
-}
+use std::collections::BTreeMap;
 
 #[test]
 fn same_package_child_does_not_shadow_inherited_parent_and_bubbles_by_name() {
@@ -498,74 +371,6 @@ fn resolved_peer_providers_from_direct_outputs_are_last_write_wins() {
     let result = resolve_peers(&mut tree, ResolvePeersOptions::default());
 
     assert_eq!(result.resolved_peer_providers_by_alias.get("peer"), Some(&second_peer));
-}
-
-/// See [`PeersCacheItem`] for why a cache hit reports no providers.
-#[test]
-fn cached_subtree_reuse_reports_no_peer_providers() {
-    let peerx = NodeId::leaf("peerx@1.0.0");
-    let peerpkg = NodeId::leaf("peerpkg@2.0.0");
-    let consumer = NodeId::next();
-    let mid = NodeId::next();
-
-    let mut mid_children = BTreeMap::new();
-    mid_children.insert("peerpkg".to_string(), peerpkg.clone());
-    mid_children.insert("consumer".to_string(), consumer.clone());
-
-    let mut tree = ResolvedTree {
-        direct: vec![
-            DirectDep {
-                alias: "mid".to_string(),
-                node_id: mid.clone(),
-                id: "mid@1.0.0".to_string(),
-            },
-            DirectDep {
-                alias: "peerx".to_string(),
-                node_id: peerx.clone(),
-                id: "peerx@1.0.0".to_string(),
-            },
-        ],
-        packages: HashMap::from_iter([
-            ("peerx@1.0.0".to_string(), package("peerx", "1.0.0", &[], true)),
-            ("peerpkg@2.0.0".to_string(), package("peerpkg", "2.0.0", &[], true)),
-            (
-                "consumer@1.0.0".to_string(),
-                package("consumer", "1.0.0", &[("peerpkg", "*"), ("peerx", "*")], false),
-            ),
-            ("mid@1.0.0".to_string(), package("mid", "1.0.0", &[], false)),
-        ]),
-        dependencies_tree: HashMap::from_iter([
-            (peerx, tree_node("peerx@1.0.0", BTreeMap::new(), 0)),
-            (peerpkg, tree_node("peerpkg@2.0.0", BTreeMap::new(), 1)),
-            (consumer, tree_node("consumer@1.0.0", BTreeMap::new(), 1)),
-            (mid, tree_node("mid@1.0.0", mid_children, 0)),
-        ]),
-        all_peer_dep_names: HashSet::from_iter(["peerpkg".to_string(), "peerx".to_string()]),
-        policy_violations: Vec::new(),
-        applied_patches: HashSet::default(),
-        children_by_id: HashMap::default(),
-    };
-    let direct = tree.direct.clone();
-
-    let (first, caches) = super::discover_peers(
-        &mut tree,
-        &direct,
-        &direct,
-        PeerDiscoveryCaches::default(),
-        ResolvePeersOptions::default(),
-    );
-    assert!(
-        first.resolved_peer_providers_by_alias.contains_key("peerpkg"),
-        "the walk that resolves the subtree reports its providers",
-    );
-
-    let (second, _) =
-        super::discover_peers(&mut tree, &direct, &direct, caches, ResolvePeersOptions::default());
-    assert_eq!(
-        second.resolved_peer_providers_by_alias.get("peerpkg"),
-        None,
-        "a cached-subtree reuse must not re-report the owner walk's providers",
-    );
 }
 
 #[test]
@@ -955,422 +760,6 @@ fn same_package_child_replaces_inherited_parent_when_peer_diamond_conflicts() {
         "plugin must not mix the root parser context with nested ts: {:#?}",
         result.graph.keys().collect::<Vec<_>>(),
     );
-}
-
-#[test]
-fn previously_resolved_children_prefers_closest_same_package_ancestor() {
-    let far_parent = NodeId::next();
-    let close_parent = NodeId::next();
-    let far_child = NodeId::leaf("shared@1.0.0");
-    let close_child = NodeId::leaf("shared@2.0.0");
-
-    let mut far_children = BTreeMap::new();
-    far_children.insert("shared".to_string(), far_child);
-    let mut close_children = BTreeMap::new();
-    close_children.insert("shared".to_string(), close_child.clone());
-
-    let mut tree = ResolvedTree {
-        direct: Vec::new(),
-        packages: HashMap::from_iter([(
-            "loop@1.0.0".to_string(),
-            package("loop", "1.0.0", &[], false),
-        )]),
-        dependencies_tree: HashMap::from_iter([
-            (far_parent.clone(), tree_node("loop@1.0.0", far_children, 0)),
-            (close_parent.clone(), tree_node("loop@1.0.0", close_children, 2)),
-        ]),
-        all_peer_dep_names: HashSet::default(),
-        policy_violations: Vec::new(),
-        applied_patches: HashSet::default(),
-        children_by_id: HashMap::default(),
-    };
-    let mut walker = walker_for_tests(&mut tree);
-
-    let parent_node_ids = super::SharedChain::default().pushed(far_parent).pushed(close_parent);
-    let parent_pkg_ids = super::SharedChain::default().pushed("loop@1.0.0".to_string());
-    let children =
-        walker.previously_resolved_children(&parent_node_ids, &parent_pkg_ids, "loop@1.0.0");
-
-    assert_eq!(children.get("shared"), Some(&close_child));
-}
-
-#[test]
-fn final_graph_keeps_first_equal_depth_payload_and_unions_transitive_peers() {
-    let first = NodeId::next();
-    let second = NodeId::next();
-    let first_child = NodeId::leaf("child-a@1.0.0");
-    let second_child = NodeId::leaf("child-b@1.0.0");
-    let final_dep_path = DepPath::from("same@1.0.0(peer@1.0.0)");
-
-    let mut tree = ResolvedTree {
-        direct: Vec::new(),
-        packages: HashMap::from_iter([
-            ("same@1.0.0".to_string(), package("same", "1.0.0", &[("peer", "*")], false)),
-            ("child-a@1.0.0".to_string(), package("child-a", "1.0.0", &[], true)),
-            ("child-b@1.0.0".to_string(), package("child-b", "1.0.0", &[], true)),
-        ]),
-        dependencies_tree: HashMap::from_iter([
-            (first.clone(), tree_node("same@1.0.0", BTreeMap::new(), 1)),
-            (second.clone(), tree_node("same@1.0.0", BTreeMap::new(), 1)),
-            (first_child.clone(), tree_node("child-a@1.0.0", BTreeMap::new(), 2)),
-            (second_child.clone(), tree_node("child-b@1.0.0", BTreeMap::new(), 2)),
-        ]),
-        all_peer_dep_names: HashSet::from_iter(["peer".to_string()]),
-        policy_violations: Vec::new(),
-        applied_patches: HashSet::default(),
-        children_by_id: HashMap::default(),
-    };
-    let mut walker = walker_for_tests(&mut tree);
-    walker.node_dep_paths.insert(first.clone(), final_dep_path.clone());
-    walker.node_dep_paths.insert(second.clone(), final_dep_path.clone());
-    walker.node_dep_paths.insert(first_child.clone(), DepPath::from("child-a@1.0.0"));
-    walker.node_dep_paths.insert(second_child.clone(), DepPath::from("child-b@1.0.0"));
-    walker.node_records.insert(
-        second.clone(),
-        NodeRecord {
-            edges: BTreeMap::from([("peer".to_string(), second_child)]),
-            optional_child_aliases: HashSet::default(),
-            transitive_peer_dependencies: HashSet::from_iter(["debug".to_string()]),
-            depth: 1,
-            installable: true,
-            is_pure: false,
-            order: 1,
-        },
-    );
-    walker.node_records.insert(
-        first.clone(),
-        NodeRecord {
-            edges: BTreeMap::from([("peer".to_string(), first_child)]),
-            optional_child_aliases: HashSet::default(),
-            transitive_peer_dependencies: HashSet::default(),
-            depth: 1,
-            installable: true,
-            is_pure: false,
-            order: 0,
-        },
-    );
-
-    let graph = walker.build_final_graph(&HashMap::from_iter([
-        (first, final_dep_path.clone()),
-        (second, final_dep_path.clone()),
-    ]));
-
-    assert_eq!(graph[&final_dep_path].children.get("peer"), Some(&DepPath::from("child-a@1.0.0")));
-    assert!(graph[&final_dep_path].transitive_peer_dependencies.contains("debug"));
-}
-
-#[test]
-fn final_graph_duplicate_parent_prefers_child_variant_matching_parent_peers() {
-    let first = NodeId::next();
-    let second = NodeId::next();
-    let first_child = NodeId::next();
-    let second_child = NodeId::next();
-    let parent_dep_path = DepPath::from(
-        "consumer@1.0.0(webpack-cli@6.0.1)(webpack-dev-server@5.2.2)(webpack@5.107.2)",
-    );
-    let analyzer_child_dep_path = DepPath::from(
-        "webpack-cli@6.0.1(webpack-bundle-analyzer@4.10.2)(webpack-dev-server@5.2.2)(webpack@5.107.2)",
-    );
-    let matching_child_dep_path =
-        DepPath::from("webpack-cli@6.0.1(webpack-dev-server@5.2.2)(webpack@5.107.2)");
-
-    let mut tree = ResolvedTree {
-        direct: Vec::new(),
-        packages: HashMap::from_iter([(
-            "consumer@1.0.0".to_string(),
-            package("consumer", "1.0.0", &[], false),
-        )]),
-        dependencies_tree: HashMap::from_iter([
-            (first.clone(), tree_node("consumer@1.0.0", BTreeMap::new(), 1)),
-            (second.clone(), tree_node("consumer@1.0.0", BTreeMap::new(), 1)),
-        ]),
-        all_peer_dep_names: HashSet::default(),
-        policy_violations: Vec::new(),
-        applied_patches: HashSet::default(),
-        children_by_id: HashMap::default(),
-    };
-    let mut walker = walker_for_tests(&mut tree);
-    walker.node_dep_paths.insert(first.clone(), parent_dep_path.clone());
-    walker.node_dep_paths.insert(second.clone(), parent_dep_path.clone());
-    walker.node_records.insert(
-        first.clone(),
-        NodeRecord {
-            edges: BTreeMap::from([("webpack-cli".to_string(), first_child.clone())]),
-            optional_child_aliases: HashSet::default(),
-            transitive_peer_dependencies: HashSet::default(),
-            depth: 1,
-            installable: true,
-            is_pure: false,
-            order: 0,
-        },
-    );
-    walker.node_records.insert(
-        second.clone(),
-        NodeRecord {
-            edges: BTreeMap::from([("webpack-cli".to_string(), second_child.clone())]),
-            optional_child_aliases: HashSet::default(),
-            transitive_peer_dependencies: HashSet::default(),
-            depth: 1,
-            installable: true,
-            is_pure: false,
-            order: 1,
-        },
-    );
-
-    let graph = walker.build_final_graph(&HashMap::from_iter([
-        (first, parent_dep_path.clone()),
-        (second, parent_dep_path.clone()),
-        (first_child, analyzer_child_dep_path),
-        (second_child, matching_child_dep_path.clone()),
-    ]));
-
-    assert_eq!(graph[&parent_dep_path].children.get("webpack-cli"), Some(&matching_child_dep_path));
-}
-
-#[test]
-fn final_graph_peer_edge_keeps_the_providers_own_peer_suffix() {
-    let provider_analyzer = NodeId::next();
-    let provider_bare = NodeId::next();
-    let consumer = NodeId::next();
-
-    let provider_analyzer_dep_path = DepPath::from(
-        "webpack-cli@6.0.1(webpack-bundle-analyzer@4.10.2)(webpack-dev-server@5.2.2)(webpack@5.107.2)",
-    );
-    let trimmed_dep_path =
-        DepPath::from("webpack-cli@6.0.1(webpack-dev-server@5.2.2)(webpack@5.107.2)");
-    let provider_bare_dep_path = DepPath::from("webpack-cli@6.0.1(webpack@5.107.2)");
-    let consumer_dep_path = DepPath::from(
-        "@webpack-cli/serve@3.0.1(webpack-cli@6.0.1)(webpack-dev-server@5.2.2)(webpack@5.107.2)",
-    );
-
-    let mut tree = ResolvedTree {
-        direct: Vec::new(),
-        packages: HashMap::from_iter([
-            (
-                "webpack-cli@6.0.1".to_string(),
-                package(
-                    "webpack-cli",
-                    "6.0.1",
-                    &[
-                        ("webpack", "*"),
-                        ("webpack-dev-server", "*"),
-                        ("webpack-bundle-analyzer", "*"),
-                    ],
-                    false,
-                ),
-            ),
-            (
-                "@webpack-cli/serve@3.0.1".to_string(),
-                package(
-                    "@webpack-cli/serve",
-                    "3.0.1",
-                    &[("webpack", "*"), ("webpack-cli", "*"), ("webpack-dev-server", "*")],
-                    false,
-                ),
-            ),
-        ]),
-        dependencies_tree: HashMap::from_iter([
-            (provider_analyzer.clone(), tree_node("webpack-cli@6.0.1", BTreeMap::new(), 0)),
-            (provider_bare.clone(), tree_node("webpack-cli@6.0.1", BTreeMap::new(), 0)),
-            (consumer.clone(), tree_node("@webpack-cli/serve@3.0.1", BTreeMap::new(), 1)),
-        ]),
-        all_peer_dep_names: HashSet::from_iter([
-            "webpack".to_string(),
-            "webpack-cli".to_string(),
-            "webpack-dev-server".to_string(),
-            "webpack-bundle-analyzer".to_string(),
-        ]),
-        policy_violations: Vec::new(),
-        applied_patches: HashSet::default(),
-        children_by_id: HashMap::default(),
-    };
-    let mut walker = walker_for_tests(&mut tree);
-    walker.node_records.insert(
-        provider_analyzer.clone(),
-        NodeRecord {
-            edges: BTreeMap::new(),
-            optional_child_aliases: HashSet::default(),
-            transitive_peer_dependencies: HashSet::default(),
-            depth: 0,
-            installable: true,
-            is_pure: false,
-            order: 0,
-        },
-    );
-    walker.node_records.insert(
-        provider_bare.clone(),
-        NodeRecord {
-            edges: BTreeMap::new(),
-            optional_child_aliases: HashSet::default(),
-            transitive_peer_dependencies: HashSet::default(),
-            depth: 0,
-            installable: true,
-            is_pure: false,
-            order: 1,
-        },
-    );
-    walker.node_records.insert(
-        consumer.clone(),
-        NodeRecord {
-            edges: BTreeMap::from([("webpack-cli".to_string(), provider_analyzer.clone())]),
-            optional_child_aliases: HashSet::default(),
-            transitive_peer_dependencies: HashSet::default(),
-            depth: 1,
-            installable: true,
-            is_pure: false,
-            order: 2,
-        },
-    );
-    walker.node_external_peers.insert(
-        provider_analyzer.clone(),
-        Arc::new(HashMap::from_iter([
-            ("webpack-bundle-analyzer".to_string(), NodeId::leaf("webpack-bundle-analyzer@4.10.2")),
-            ("webpack-dev-server".to_string(), NodeId::leaf("webpack-dev-server@5.2.2")),
-            ("webpack".to_string(), NodeId::leaf("webpack@5.107.2")),
-        ])),
-    );
-    walker.node_external_peers.insert(
-        provider_bare.clone(),
-        Arc::new(HashMap::from_iter([("webpack".to_string(), NodeId::leaf("webpack@5.107.2"))])),
-    );
-    walker.node_external_peers.insert(
-        consumer.clone(),
-        Arc::new(HashMap::from_iter([
-            ("webpack-bundle-analyzer".to_string(), NodeId::leaf("webpack-bundle-analyzer@4.10.2")),
-            ("webpack-cli".to_string(), provider_analyzer.clone()),
-            ("webpack-dev-server".to_string(), NodeId::leaf("webpack-dev-server@5.2.2")),
-            ("webpack".to_string(), NodeId::leaf("webpack@5.107.2")),
-        ])),
-    );
-
-    let graph = walker.build_final_graph(&HashMap::from_iter([
-        (provider_analyzer, provider_analyzer_dep_path.clone()),
-        (provider_bare, provider_bare_dep_path),
-        (consumer, consumer_dep_path.clone()),
-    ]));
-
-    assert_eq!(
-        graph[&consumer_dep_path].children.get("webpack-cli"),
-        Some(&provider_analyzer_dep_path),
-    );
-    assert!(!graph.contains_key(&trimmed_dep_path), "no variant is fabricated for the edge");
-}
-
-#[test]
-fn final_graph_peer_edge_keeps_provider_transitive_peer_suffixes() {
-    let provider = NodeId::next();
-    let consumer = NodeId::next();
-
-    let provider_dep_path = DepPath::from(
-        "webpack-dev-server@5.2.2(bufferutil@4.1.0)(tslib@2.8.1)(utf-8-validate@5.0.10)(webpack-cli@6.0.1)(webpack@5.107.2)",
-    );
-    let trimmed_provider_dep_path =
-        DepPath::from("webpack-dev-server@5.2.2(webpack-cli@6.0.1)(webpack@5.107.2)");
-    let consumer_dep_path = DepPath::from(
-        "webpack-cli@6.0.1(webpack-bundle-analyzer@4.10.2)(webpack-dev-server@5.2.2)(webpack@5.107.2)",
-    );
-
-    let mut tree = ResolvedTree {
-        direct: Vec::new(),
-        packages: HashMap::from_iter([
-            (
-                "webpack-dev-server@5.2.2".to_string(),
-                package(
-                    "webpack-dev-server",
-                    "5.2.2",
-                    &[("webpack", "*"), ("webpack-cli", "*")],
-                    false,
-                ),
-            ),
-            (
-                "webpack-cli@6.0.1".to_string(),
-                package(
-                    "webpack-cli",
-                    "6.0.1",
-                    &[
-                        ("webpack", "*"),
-                        ("webpack-bundle-analyzer", "*"),
-                        ("webpack-dev-server", "*"),
-                    ],
-                    false,
-                ),
-            ),
-        ]),
-        dependencies_tree: HashMap::from_iter([
-            (provider.clone(), tree_node("webpack-dev-server@5.2.2", BTreeMap::new(), 1)),
-            (consumer.clone(), tree_node("webpack-cli@6.0.1", BTreeMap::new(), 0)),
-        ]),
-        all_peer_dep_names: HashSet::from_iter([
-            "bufferutil".to_string(),
-            "tslib".to_string(),
-            "utf-8-validate".to_string(),
-            "webpack".to_string(),
-            "webpack-cli".to_string(),
-            "webpack-dev-server".to_string(),
-            "webpack-bundle-analyzer".to_string(),
-        ]),
-        policy_violations: Vec::new(),
-        applied_patches: HashSet::default(),
-        children_by_id: HashMap::default(),
-    };
-    let mut walker = walker_for_tests(&mut tree);
-    walker.node_records.insert(
-        provider.clone(),
-        NodeRecord {
-            edges: BTreeMap::new(),
-            optional_child_aliases: HashSet::default(),
-            transitive_peer_dependencies: HashSet::from_iter([
-                "bufferutil".to_string(),
-                "tslib".to_string(),
-                "utf-8-validate".to_string(),
-            ]),
-            depth: 1,
-            installable: true,
-            is_pure: false,
-            order: 0,
-        },
-    );
-    walker.node_records.insert(
-        consumer.clone(),
-        NodeRecord {
-            edges: BTreeMap::from([("webpack-dev-server".to_string(), provider.clone())]),
-            optional_child_aliases: HashSet::default(),
-            transitive_peer_dependencies: HashSet::default(),
-            depth: 0,
-            installable: true,
-            is_pure: false,
-            order: 1,
-        },
-    );
-    walker.node_external_peers.insert(
-        provider.clone(),
-        Arc::new(HashMap::from_iter([
-            ("bufferutil".to_string(), NodeId::leaf("bufferutil@4.1.0")),
-            ("tslib".to_string(), NodeId::leaf("tslib@2.8.1")),
-            ("utf-8-validate".to_string(), NodeId::leaf("utf-8-validate@5.0.10")),
-            ("webpack-cli".to_string(), consumer.clone()),
-            ("webpack".to_string(), NodeId::leaf("webpack@5.107.2")),
-        ])),
-    );
-    walker.node_external_peers.insert(
-        consumer.clone(),
-        Arc::new(HashMap::from_iter([
-            ("webpack-bundle-analyzer".to_string(), NodeId::leaf("webpack-bundle-analyzer@4.10.2")),
-            ("webpack-dev-server".to_string(), provider.clone()),
-            ("webpack".to_string(), NodeId::leaf("webpack@5.107.2")),
-        ])),
-    );
-
-    let graph = walker.build_final_graph(&HashMap::from_iter([
-        (provider, provider_dep_path.clone()),
-        (consumer, consumer_dep_path.clone()),
-    ]));
-
-    assert_eq!(
-        graph[&consumer_dep_path].children.get("webpack-dev-server"),
-        Some(&provider_dep_path),
-    );
-    assert!(!graph.contains_key(&trimmed_provider_dep_path));
 }
 
 // Parity check for <https://github.com/pnpm/pnpm/pull/12514>.
@@ -2139,138 +1528,12 @@ fn peer_cycle_between_own_dep_and_provider_at_tree_position_resolves() {
     );
 }
 
-fn tree_node(pkg_id: &str, children: BTreeMap<String, NodeId>, depth: i32) -> DependenciesTreeNode {
-    DependenciesTreeNode::new(pkg_id.to_string(), TreeChildren::Realized(children), depth, true)
-}
-
-#[test]
-fn discovery_engine_rebuilds_after_a_children_ownership_rewrite() {
-    use super::PeerHoistDiscovery;
-    use crate::resolve_dependency_tree::WorkspaceTreeCtx;
-
-    let workspace = WorkspaceTreeCtx::default();
-    let mut engine = PeerHoistDiscovery::new();
-    engine.discover(&workspace, &[], &[], ResolvePeersOptions::default());
-
-    // A marker in the persistent caches makes reset-vs-merge observable.
-    engine.caches.pure_pkgs.insert("marker@1.0.0".to_string(), DepPath::from("marker@1.0.0"));
-    // Production rewrites happen inside `extend_tree`, which always
-    // bumps the revision; mirror that pairing.
-    workspace.record_children_rewrite();
-    workspace.bump_revision();
-    engine.discover(&workspace, &[], &[], ResolvePeersOptions::default());
-    assert!(
-        engine.caches.pure_pkgs.is_empty(),
-        "an ownership rewrite must discard walk state derived before it",
-    );
-
-    engine.caches.pure_pkgs.insert("marker@1.0.0".to_string(), DepPath::from("marker@1.0.0"));
-    workspace.bump_revision();
-    engine.discover(&workspace, &[], &[], ResolvePeersOptions::default());
-    assert!(
-        engine.caches.pure_pkgs.contains_key("marker@1.0.0"),
-        "a rewrite-free revision bump merges instead of rebuilding",
-    );
-}
-
-fn walker_for_tests(tree: &mut ResolvedTree) -> Walker<'_> {
-    Walker::new(
-        tree,
-        ResolvePeersOptions::default(),
-        HashMap::default(),
-        Vec::new(),
-        PeerDiscoveryCaches::default(),
-        false,
-    )
-}
-
-fn package(
-    name: &str,
-    version: &str,
-    peer_dependencies: &[(&str, &str)],
-    is_leaf: bool,
-) -> ResolvedPackage {
-    let peer_dependencies: Vec<_> =
-        peer_dependencies.iter().map(|(name, version)| (*name, *version, false)).collect();
-    package_with_peer_dependencies(name, version, &peer_dependencies, is_leaf)
-}
-
-fn package_with_peer_dependencies(
-    name: &str,
-    version: &str,
-    peer_dependencies: &[(&str, &str, bool)],
-    is_leaf: bool,
-) -> ResolvedPackage {
-    let peer_dependencies = peer_dependencies
-        .iter()
-        .map(|(name, version, optional)| {
-            ((*name).to_string(), PeerDep { version: (*version).to_string(), optional: *optional })
-        })
-        .collect();
-    ResolvedPackage {
-        id: format!("{name}@{version}"),
-        result: Arc::new(resolve_result(name, version)),
-        peer_dependencies,
-        optional: false,
-        is_leaf,
-    }
-}
-
-fn linked_package(name: &str, id: &str, directory: &str) -> ResolvedPackage {
-    ResolvedPackage {
-        id: id.to_string(),
-        result: Arc::new(ResolveResult {
-            id: PkgResolutionId::from(id.to_string()),
-            name_ver: None,
-            latest: None,
-            published_at: None,
-            manifest: Some(Arc::new(serde_json::json!({ "name": name, "version": "1.0.0" }))),
-            resolution: LockfileResolution::Directory(DirectoryResolution {
-                directory: directory.to_string(),
-            }),
-            resolved_via: "workspace".to_string(),
-            normalized_bare_specifier: None,
-            alias: Some(name.to_string()),
-            policy_violation: None,
-        }),
-        peer_dependencies: BTreeMap::new(),
-        optional: false,
-        is_leaf: true,
-    }
-}
-
-fn resolve_result(name: &str, version: &str) -> ResolveResult {
-    let name_ver = PkgNameVer::new(
-        PkgName::parse(name).unwrap(),
-        node_semver::Version::from_str(version).unwrap(),
-    );
-    ResolveResult {
-        id: (&name_ver).into(),
-        name_ver: Some(name_ver),
-        latest: Some(version.to_string()),
-        published_at: None,
-        manifest: None,
-        resolution: LockfileResolution::Tarball(TarballResolution {
-            tarball: format!("https://registry.example/{name}-{version}.tgz"),
-            integrity: None,
-            git_hosted: None,
-            path: None,
-        }),
-        resolved_via: "npm-registry".to_string(),
-        normalized_bare_specifier: None,
-        alias: Some(name.to_string()),
-        policy_violation: None,
-    }
-}
-
 /// Ported from upstream `resolvePeers.ts`'s `locked peer provider
 /// preferences` suite: a second resolution pass receives the first
 /// pass's `paths_by_node_id` and re-pins compatible locked providers.
 mod locked_peer_provider_preferences {
-    use super::{
-        DepPath, DirectDep, NodeId, ResolvePeersOptions, ResolvedTree, package,
-        package_with_peer_dependencies, resolve_peers, tree_node,
-    };
+    use super::{DepPath, DirectDep, NodeId, ResolvePeersOptions, ResolvedTree, resolve_peers};
+    use crate::resolve_peers::test_support::{package, package_with_peer_dependencies, tree_node};
     use rustc_hash::{FxHashMap as HashMap, FxHashSet as HashSet};
     use std::collections::BTreeMap;
